% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caBrush.R
\name{caBrush}
\alias{caBrush}
\title{Cleaning cantaloupes by brushing}
\usage{
caBrush(data = list(), nLots = NULL, sizeLot = NULL, logDecBrush)
}
\arguments{
\item{data}{a list of:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes} on the rind,
from contaminated cultivation lots;}
\item{\code{P}}{Prevalence of contaminated harvested lots (scalar).}
}}

\item{nLots}{see \code{\link[=Lot2LotGen]{Lot2LotGen()}} function.}

\item{sizeLot}{see \code{\link[=Lot2LotGen]{Lot2LotGen()}} function.}

\item{logDecBrush}{Mean log10 reduction attained by brushing or scrubbing (scalar, vector of size the number of \code{N} rows
(representing lot-to-lot variability), or vector of size the number of \code{N} columns (representing element-to-element variability)).}
}
\value{
A list of two elements of the data objects:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units containing the numbers of \emph{L. monocytogenes}
on brushed cantaloupes}
\item{\code{P}}{Prevalence of contaminated harvested lots post-brushing (scalar).}
}
}
\description{
The function \code{\link[=caBrush]{caBrush()}} models the removal of bacteria during the brushing or scrubbing step.
The mean log10 reduction due to brushing must be provided.
}
\note{
Abrasive brushing for cantaloupe cleaning has been shown to be efficient in removing biofilms of \emph{L. monocytogenes}
attached to the rind in \insertCite{Fu2020;textual}{qraLm}.
}
\examples{
sizeLot <- 100
nLots <- 500
df <- list(N=matrix(rpois(sizeLot * nLots, 10),
             nrow = nLots, ncol = sizeLot),
              P=0.7, cantaWeight=1000, sizeLot=100,
              nLots=500)
brushed <- caBrush(df,
                   logDecBrush = runif(nLots, min = 0, max = 10)
                )
hist(brushed$N)

}
\references{
\insertRef{mc2d}{qraLm}

\insertRef{stats}{qraLm}

\insertRef{Fu2020}{qraLm}
}
\author{
Regis Pouillot \email{rpouillot.work@gmail.com}
}
\keyword{mechanical}
\keyword{melon}
\keyword{removal}
\keyword{scrubbing}
