% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caDicing.R
\name{caDicing}
\alias{caDicing}
\title{Dicing of cantaloupe for preparation of RTE pre-cut cantaloupe}
\usage{
caDicing(
  data = list(),
  minTR = 0.087,
  modeTR = 0.55,
  maxTR = 2.82,
  cantaSurface = 580,
  pulpYield = 0.6,
  cantaWeight = NULL,
  nLots = NULL,
  sizeLot = NULL,
  sizeSublot = 20
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes}
on cantaloupe rind, from contaminated lots;}
\item{\code{P}}{Prevalence of contaminated harvested lots (scalar).}
}}

\item{minTR}{(\verb{\%}) Minimum transfer rate from rind to flesh (\eqn{default=0.087} for use in a Pert distribution) (scalar).}

\item{modeTR}{(\verb{\%}) Mode of the transfer rate from rind to flesh (\eqn{default=0.55} for use in a Pert distribution) (scalar).}

\item{maxTR}{(\verb{\%}) Maximum transfer rate from rind to flesh (\eqn{default=2.82} for use in a Pert distribution) (scalar).}

\item{cantaSurface}{(\eqn{cm^2}) Surface area of a cantaloupe (scalar or vector).}

\item{pulpYield}{(\verb{\%}) Cantaloupe pulp yield (scalar or vector).}

\item{cantaWeight}{(\code{g}) weight of a cantaloupe}

\item{nLots}{Number of lots (scalar).}

\item{sizeLot}{Number of cantaloupes in a production lot (scalar).}

\item{sizeSublot}{Number of cantaloupes to be diced in a sublot or processing lot (the sizeSublot should be multiple of sizeLot) (scalar).}
}
\value{
A list of two elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix in the sublot arrangement size \code{newMC} sublots by sublot_size portions of dices from one cantaloupe
representing the numbers of \emph{L. monocytogenes} in dices, from contaminated production lots.}
\item{\code{P}}{Prevalence of sublots of dices contaminated with \emph{L. monocytogenes}(scalar).}
}
}
\description{
The function \code{\link[=caDicing]{caDicing()}} simulates the transfer of \emph{L. monocytogenes} from rind to flesh during dicing
of cantaloupe at processing environment. It is assumed that every cantaloupe is separately diced,
and if contaminated on the rind, a fraction of \emph{L. monocytogenes} cells is transferred to the dices.
Notice that this function does not consider cross-contamination from dicing machine or from knives.
The transfer rate (\code{TR}) in \% (\code{CFU/g} dices per \eqn{CFU/cm^2\ rind)*100} has been modelled as a Pert distribution
with parameters \code{minTR}, \code{modeTR} and \code{maxTR}.
}
\note{
The matrix \code{N} is returned in a sublot arrangement, meaning that the number of rows corresponds to the number of sublots,
and the number of columns the number of cantaloupes processed in a sublot. This is done because the production lot size is not
necessarily the same as the processing lot size. Therefore, the numbers and prevalence outputs are given in the sublot basis (processing lot).
The variability in the transfer rate from rind to flesh (\code{CFU/g} dices per \eqn{CFU/cm^2} rind) was modelled as a Pert distribution using data
digitised from \insertCite{Ukuku2002;textual}{JEMRA} and \insertCite{Ukuku2012;textual}{JEMRA}.
}
\examples{
nLots        <- 1000
sizeLot      <- 100
sizeSublot   <- 20
cantaSurface <- 780
pulpYield    <- 0.75
cantaWeight  <- 1000
data         <- list(N = matrix(rpois(sizeLot*nLots, 100),
                                ncol=sizeLot, 
                                nrow=nLots),
                                P = 0.16,
                      cantaWeight=cantaWeight,
                      nLots=nLots,
                      sizeLot=sizeLot)
str(data)
CantaDices <- caDicing (data,
                        minTR=0.087,
                        modeTR=0.55,
                        maxTR=2.82,
                        pulpYield=pulpYield,
                        cantaSurface=cantaSurface,
                        sizeSublot=sizeSublot)
hist(CantaDices$N)

}
\references{
\insertRef{mc2d}{JEMRA}

\insertRef{Ukuku2002}{JEMRA}

\insertRef{Ukuku2012}{JEMRA}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{cutting}
\keyword{flesh}
\keyword{melon}
\keyword{rind}
\keyword{transfer}
