% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caFlumeTank.R
\name{caFlumeTank}
\alias{caFlumeTank}
\title{Cantaloupe washing in flume tank}
\usage{
caFlumeTank(
  data = list(),
  logDecWash,
  logDecSani,
  b,
  nLots = NULL,
  sizeLot = NULL
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes}
on the rind, from contaminated cultivation lots;}
\item{\code{P}}{Prevalence of contaminated harvested lots pre-washing (scalar).}
}}

\item{logDecWash}{(log10) Reduction attained by washing (scalar or vector).}

\item{logDecSani}{(log10) Reduction attained by sanitising water (scalar or vector).}

\item{b}{Dispersion factor representing the clustering of surviving cells during redistribution (scalar, vector or matrix).}

\item{nLots}{Number of harvested lots of cantaloupe.}

\item{sizeLot}{Number of cantaloupes in a harvested lot.}
}
\value{
A list of two elements of the data objects:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes}
on washed cantaloupes.}
\item{\code{P}}{Prevalence of contaminated harvested lots post-washing (scalar).}
}
}
\description{
The function \code{\link[=caFlumeTank]{caFlumeTank()}} simulates three events taking place during flume tank washing:
\enumerate{
\item the washing of bacteria off the cantaloupe rind at a washing efficiency of \code{logDecWash};
\item the elimination of bacteria in the sanitising water at a sanitising efficiency of \code{log10SaniWash}; and
\item the redistribution of survivors on the cantaloupe rind.
}
}
\note{
Since surviving cells are distributed in water, a value  of \code{b} higher than 1 should be used to assume
random homogeneous distribution. This dispersion factor is a parameter of the beta-binomial
distribution (\insertCite{Nauta2005;textual}{qraLm}).
}
\examples{
library(extraDistr)
library(mc2d)
dat <- caPrimaryProduction(
  nLots = 100,
  sizeLot = 100,
  cantaWeight = 800,
  pSoil = 0.089,
  pManure = 0.1,
  pIrrigRaining = 0.05,
  pFoil = 0.5,
  rFoil = 0.75,
  pIrrig = 0.131
)

WashedMelons <- caFlumeTank(dat,
  logDecWash = 1.5,
  logDecSani = 2,
  b = 2.5,
  nLots = nLots,
  sizeLot = sizeLot
)
hist(WashedMelons$N)

}
\references{
\insertRef{mc2d}{qraLm}

\insertRef{stats}{qraLm}

\insertRef{Nauta2005}{qraLm}
}
\author{
Regis Pouillot \email{rpouillot.work@gmail.com}
}
\keyword{immersion}
\keyword{removal}
\keyword{water}
