% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caFlumeTankCC.R
\name{caFlumeTankCC}
\alias{caFlumeTankCC}
\title{Cantaloupe cross-contamination during washing in flume tank}
\usage{
caFlumeTankCC(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  cantaWeight = NULL,
  probCCW = 0.5,
  logWaterMin = 1,
  logWaterMode = 1,
  logWaterMax = 5,
  pWaterGain = 0.004,
  bWater = 1
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes}
on the rind, from contaminated cultivation lots;}
\item{\code{P}}{Prevalence of contaminated harvested lots pre-washing (scalar).}
}}

\item{nLots}{see \code{\link[=Lot2LotGen]{Lot2LotGen()}} function.}

\item{sizeLot}{see \code{\link[=Lot2LotGen]{Lot2LotGen()}} function.}

\item{cantaWeight}{(\code{g}) weight of a cantaloupe.}

\item{probCCW}{Probability that water of flume tank is contaminated  (scalar).}

\item{logWaterMin}{(log10(cfu/l)) Minimal concentration of \emph{L. monocytogenes} in water of flume tank  (scalar or vector). The value was derived during JEMRA meeting related to unpublished data transmitted by experts}

\item{logWaterMode}{(log10(cfu/l))  Mode concentration of \emph{L. monocytogenes} in water of flume tank  (scalar or vector). The value was derived during JEMRA meeting related to unpublished data transmitted by experts}

\item{logWaterMax}{(log10(cfu/l)) Maximal concentration of \emph{L. monocytogenes} in water of flume tank  (scalar or vector). The value was derived during JEMRA meeting related to unpublished data transmitted by experts}

\item{pWaterGain}{value of the fraction of water gain (ml) relative to the cantaloupe weight in g (\eqn{default=0.004} according to \insertCite{Richards2004;textual}{qraLm}).}

\item{bWater}{Dispersion factor representing the clustering of cells during flume_tank washing (scalar).}
}
\value{
A list of two elements of the data objects:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes}
on washed cantaloupes.}
\item{\code{P}}{Prevalence of contaminated harvested lots post-washing (scalar).}
}
}
\description{
The function \code{\link[=caFlumeTankCC]{caFlumeTankCC()}} function simulates the potential contamination of cantaloupe, when in direct
contact with contaminated water in flume tank.
The cross-contamination algorithm accounts for four possible scenarios:
\enumerate{
\item cross-contamination occurring in lots already contaminated;
\item re-contamination occurring in lots that were not contaminated;
\item no cross-contamination occurring in lots already contaminated; and
\item no cross-contamination occurring in lots that were not contaminated.
}
}
\note{
Write a note!
}
\examples{
dat <- caPrimaryProduction(
  nLots = 100,
    sizeLot = 100)
    Nf <- caFlumeTankCC(dat,
                   probCCW = 0.125,
                   logWaterMin = 1,
                   logWaterMode = 1,
                   logWaterMax = 5,
                   bWater = 1)
hist(Nf$N)

}
\references{
\insertRef{mc2d}{qraLm}
\insertRef{stats}{qraLm}
}
\author{
Who is the author
}
\keyword{cross-contamination}
\keyword{immersion}
\keyword{water}
