% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caHarvestCC.R
\name{caHarvestCC}
\alias{caHarvestCC}
\title{Cross-contamination of cantaloupe rind by \emph{L. monocytogenes} during harvest}
\usage{
caHarvestCC(
  data = list(),
  probCCH,
  trMean,
  trSd,
  nPlas,
  nLots = NULL,
  sizeLot = NULL
)
}
\arguments{
\item{data}{a list of
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size: \code{nLots}  by \code{sizeLot} containing the numbers of \emph{L. monocytogenes} on the rind
of cantaloupes from contaminated cultivation lots}
\item{\code{P}}{Prevalence of contaminated cultivation lots (scalar)}
}}

\item{probCCH}{Probability of cross-contamination during harvest}

\item{trMean}{Mean of the normal distribution describing the variability in the transfer coefficient from the
surface of elements to rind (scalar)}

\item{trSd}{Standard deviation of the normal distribution describing the variability in the transfer coefficient
from the surface of elements to rind (scalar)}

\item{nPlas}{(\code{CFU}) Numbers of \emph{L. monocytogenes} on food contact surfaces (conveyors or crates at harvest)
touching the cantaloupes (scalar)}

\item{nLots}{Number of cultivation lots}

\item{sizeLot}{Number of cantaloupes cultivated and harvested in a lot}
}
\value{
A list of two elements:
\describe{
\item{\code{N}}{(CFU) A matrix of size \code{nLots}  by \code{sizeLot} of \emph{L. monocytogenes} numbers on the rind of cantaloupes
from contaminated harvested lots}
\item{\code{P}}{Prevalence of harvested contaminated lots}
}
}
\description{
The function \code{\link[=caHarvestCC]{caHarvestCC()}} simulates cross-contamination that might occur at the moment of harvesting from elements such as conveyors,
crates or plastic surfaces. This function can be used for cantaloupes harvested in large farms (intended for RTE and
sold as whole cantaloupe in formal retail) and in small farms (intended for traditional markets). Parameters to be changed accordingly.
}
\note{
The algorithm evaluates the cross-contamination event at cantaloupe level (and not at cultivation lot level), assuming that
every cantaloupe has the same probability \code{probCCH} of being contaminated from conveyors, crates or other surfaces during harvesting.
}
\examples{
# Example of use in LARGE production farms
dat <- caPrimaryProduction(
nLots = 100,
  sizeLot = 100)
probCCH <- 0.005
LargeFarms <- caHarvestCC(dat,
                          probCCH = probCCH, 
                          trMean = -1.42,
                          trSd = 0.52, 
                          nPlas = 200
                          )
hist(LargeFarms$N)
probCCH <- 0.02
nPlas <- 500
SmallFarms <- caHarvestCC(dat,
                          probCCH = probCCH, trMean = -1.42,
                          trSd = 0.52, nPlas = 500,
                          nLots = 100, sizeLot = 100
                          )
                          hist(SmallFarms$N)
}
\references{
\insertRef{stats}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt} and Regis Pouillot \email{rpouillot.work@gmail.com}
}
\keyword{contamination}
\keyword{conveyors}
\keyword{crates}
\keyword{surface}
