% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caHoldingTime.R
\name{caHoldingTime}
\alias{caHoldingTime}
\title{Holding time of cantaloupes post-harvest}
\usage{
caHoldingTime(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  pCooled,
  time,
  shape = 0.6271,
  meanD410 = 1.1309,
  sdD410 = 1.770711e-06,
  meanD25 = 2.890015,
  sdD25 = 0.2288748,
  timePrevious = NULL
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{N}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes} on
the cantaloupe rind, from contaminated harvested lots;}
\item{P}{Prevalence of contaminated harvested lots (scalar).}
}}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{pCooled}{Probability that a lot of harvested cantaloupes is cooled (4-10 \eqn{^\circ C}).}

\item{time}{(\code{h}) Holding time or storage time of a lot of harvested cantaloupes.}

\item{shape}{Shape parameter of the Weibull decay model (\eqn{default = 0.6271}, obtained from modelling data from \insertCite{Nyarko2016;textual}{qraLm}).}

\item{meanD410}{(day) Mean of the time to first log10 reduction at 4-10 \eqn{^\circ C} (default = 1.1309 day,
obtained from modelling data from \insertCite{Nyarko2016;textual}{qraLm}).}

\item{sdD410}{(day) Standard deviation of the time to first log10 reduction at 4-10 \eqn{^\circ C} (\eqn{default = 1.770711e-06} day, obtained
from modelling data from \insertCite{Nyarko2016;textual}{qraLm}).}

\item{meanD25}{(day) Mean of the time to first log10 reduction at 25 \eqn{^\circ} C) (\eqn{default = 2.890015} day, obtained from modelling data
from \insertCite{Nyarko2016;textual}{qraLm}).}

\item{sdD25}{(day) Standard deviation of the time to first log10 reduction at 25 \eqn{^\circ} C (\eqn{default = 0.2288748} day, obtained
from modelling data from \insertCite{Nyarko2016;textual}{qraLm}).}

\item{timePrevious}{(\code{h}) Storage time of cantaloupes in previous holding stage(s).}
}
\value{
A list of two elements of the data objects:
\describe{
\item{N}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes} on
cantaloupe after holding \code{time};}
\item{P}{Prevalence of contaminated harvested lots after holding \code{time};}
\item{timePrevious}{(\code{h}) Total time the cantaloupes have been stored since harvesting.}
}
}
\description{
The function \code{\link[=caHoldingTime]{caHoldingTime()}} simulates the survival of \emph{L. monocytogenes} on cantaloupe rind
during post-harvest holding time or during any short storage before cantaloupes are washed in the
packinghouse. The algorithm calculates decline in  \emph{L. monocytogenes}, and not growth on surfaces,
because it is assumed that if any event that injured the rind occurred, it would have been very recent (at harvest),
and the holding time would be too short that any growth on the bruised rind would be unlikely or negligible.
The user can define a probability \code{pCooled} that the lot of just-harvested cantaloupes is kept at cold temperatures (4-10 \eqn{^\circ} C).
If \eqn{p\_Cooled=0}, the algorithm will evaluate the survival of \emph{L. monocytogenes} on cantaloupes held at ambient temperature (25 \eqn{^\circ} C).
}
\note{
Data on the survival of \emph{L. monocytogenes} on the intact rind of two varieties of cantaloupe stored at 4, 10 and 25 \eqn{^\circ} C were
extracted from \insertCite{Nyarko2016;textual}{qraLm}. A non-linear mixed model based on the two-parameter Weibull decay equation with random effects
on melon variety showed that the time to achieve the first log10 reduction (\code{D}) at 25 \eqn{^\circ} C was significantly different from the other two temperatures,
while there was not significant difference in \code{D} between the storage temperatures of 4 and 10 \eqn{^\circ C}. Therefore, the non-linear mixed model was refitted
for two temperature classes (cool 4/10 \eqn{^\circ C}) and ambient (25 \eqn{^\circ C}). The shape parameter is fixed for the two temperature classes.
Placing random effects in \code{D} with cantaloupe variety as the clustering variable, the estimates of  \code{meanD410}, \code{sdD410}, \code{meanD25} and \code{sdD25} were obtained.
The present function represents the cultivar-specific variability about \code{D} at cool or ambient temperature as a normal distribution;
and a \code{D} value is sampled for every lot, assuming a different cantaloupe cultivar.
It is assumed that after the holding time, it is not possible that a contaminated lot become free of \emph{L. monocytogenes}. Therefore,
if by chance the rinds of all cantaloupes from a contaminated lot end up having zero counts, one cell will be assigned to one cantaloupe of the lot.
}
\examples{
dat <- caPrimaryProduction(
  nLots = 100,
    sizeLot = 100)
pCooled <- 0.75
time <- 4 # hours
AfterStorage <- caHoldingTime(dat,
  pCooled = pCooled,
  time = time,
  shape = 0.6271,
  meanD410 = 1.1309,
  sdD410 = 1.770711e-06,
  meanD25 = 2.890015,
  sdD25 = 0.2288748
)
hist(AfterStorage$N, probability = TRUE)
lines(density(AfterStorage$N), col = "black", lwd = 2)
}
\references{
\insertRef{extraDistr}{qraLm}

\insertRef{stats}{qraLm}

\insertRef{Nyarko2016}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{cantaloupe}
\keyword{decline}
\keyword{rind}
\keyword{storage}
