% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caHomeRTE.R
\name{caHomeRTE}
\alias{caHomeRTE}
\title{Storage of RTE diced cantaloupe packs at home}
\usage{
caHomeRTE(
  data = list(),
  MPD = NULL,
  Tmin = -2.0196,
  tempMin,
  tempMode,
  tempMax,
  timeMin,
  timeMode,
  timeMax
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{N}{(\code{CFU}) A matrix containing the numbers of \emph{L. monocytogenes} in packs of RTE diced cantaloupe at the end of
transport to retail, from contaminated lots;}
\item{P}{Prevalence of contaminated lots (scalar);}
\item{lnQt}{Natural log of the \code{Q} parameter at the end of transport to retail (matrix);}
\item{lotEGR5}{(\eqn{h^{-1}}) Growth rate of \emph{L. monocytogenes} in cantaloupe flesh specific to every lot (vector);}
\item{unitSize}{(\code{g}) Weight of a pack of cantaloupe dices.}
}}

\item{MPD}{(log10 CFU/g) Maximum population density of \emph{L. monocytogenes} in cantaloupe flesh (scalar).}

\item{Tmin}{(\eqn{^\circ C}) Nominal minimum temperature for growth of \emph{L. monocytogenes} in
cantaloupe flesh (suggested \eqn{value=-2.0196\ ^\circ C}) (scalar).}

\item{tempMin}{(\eqn{^\circ C}) Minimum retail temperature (scalar) (suggested \eqn{value=3.10\ ^\circ C}) (scalar).}

\item{tempMode}{(\eqn{^\circ C}) Mode of the retail temperature (scalar) (suggested \eqn{value=6.64\ ^\circ C}) (scalar).}

\item{tempMax}{(\eqn{^\circ} C) Maximum retail temperature (scalar) (suggested \eqn{value=11.3\ ^\circ C}) (scalar).}

\item{timeMin}{(\code{h}) Minimum retail time (scalar).}

\item{timeMode}{(\code{h}) Mode of the retail time (scalar).}

\item{timeMax}{(\code{h}) Maximum retail time (scalar).}
}
\value{
A list of two elements:
\describe{
\item{N}{(\code{CFU}) A matrix containing the numbers of \emph{L. monocytogenes} in packs of RTE diced
cantaloupe at the point of consumption, from contaminated lots;}
\item{P}{Prevalence of RTE diced cantaloupe lots contaminated with \emph{L. monocytogenes} (scalar).}
}
}
\description{
The function \code{\link[=caHomeRTE]{caHomeRTE()}} simulates the growth of \emph{L. monocytogenes} in RTE diced cantaloupe at home,
and is based on the function  \code{\link[=caGrowthBaranyi]{caGrowthBaranyi()}}. The algorithm samples home storage time and temperature
at the unit level since they depend on the consumer. The input \code{data} provides the algorithm with the lot-specific values of \code{EGR5}
and the unit-specific values of \code{lnQt} obtained from the previous logistics stage. Pert distributions represent the variability
in home storage time and temperature.
}
\note{
The suggested parameters for the Pert distribution of retail temperature are taken from \insertCite{Carrasco2010;textual}{qraLm},
\insertCite{Ding2013;textual}{qraLm} and \insertCite{Nauta2003;textual}{qraLm}: \eqn{tempMin=3.10\ ^\circ C}, \eqn{Temp\_mode=6.64\ ^\circ C}
and \eqn{Temp\_max=11.3\ ^\circ C}.
The parameter \eqn{Tmin=-2.0196\ ^\circ C} was determined from fitting a square-root model to data extracted from multiple sources
(refer to the function \code{\link[=caGrowthBaranyi]{caGrowthBaranyi()}}).
Parameters for the home storage time distribution should be defined by the user and/or tested in scenarios.
}
\examples{
timeMin <- 0.25
timeMode <- 3
timeMax <- 24
nLots <- 1000
sizeLot <- 250
N <- matrix(320, nLots, sizeLot)
N[2, ] <- 0
lnQt <- matrix(3.5, nLots, sizeLot)
lnQt[2, ] <- 0
data <- list(
  N = N, lnQt = lnQt,
  lotEGR5 = extraDistr::rtnorm(nLots, 0.03557288, 0.004, a = 0),
  P = 0.4, unitSize = 200
)
AtConsumption <- caHomeRTE(data,
  MPD = 8.5, Tmin = -2.0196,
  tempMin = 3.1, tempMode = 6.64, tempMax = 11.3,
  timeMin, timeMode, timeMax
)
hist(AtConsumption$N)
}
\references{
\insertRef{mc2d}{qraLm}
\insertRef{Carrasco2010}{qraLm}
\insertRef{Ding2013}{qraLm}
\insertRef{Nauta2003}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Baranyi}
\keyword{growth}
\keyword{lag-phase}
\keyword{logistics}
\keyword{stochastic}
