% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caIrrig2rind.R
\name{caIrrig2rind}
\alias{caIrrig2rind}
\title{Route of contamination of cantaloupe from irrigation water to rind}
\usage{
caIrrig2rind(
  nLots,
  sizeLot,
  pIrrig = 0.131,
  cIrrigLogMin = -1.52,
  cIrrigLogMax = 1.04,
  cantaWeight = 1000,
  pWaterGainMin = 0,
  pWaterGainMax = 0.004
)
}
\arguments{
\item{nLots}{Size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of cantaloupes per cultivation lot or field (scalar).}

\item{pIrrig}{Prevalence of contamination in irrigation water (provided by the user, \eqn{default=0.131} according to \insertCite{Raschle2021;textual}{qraLm}).}

\item{cIrrigLogMin}{(log CFU/L) Minimum value of the uniform distribution (provided by the user, \eqn{default=-1.52} according to \insertCite{Sharma2020;textual}{qraLm}).}

\item{cIrrigLogMax}{(log CFU/L)  Maximum value of the uniform distribution (provided by the user, \eqn{default=1.04} according to \insertCite{Sharma2020;textual}{qraLm}).}

\item{cantaWeight}{(g) Weight of a cantaloupe.}

\item{pWaterGainMin}{Minimum value of the fraction of water gain (ml) relative to the cantaloupe weight in g (\eqn{default=0.0}).}

\item{pWaterGainMax}{Maximum value of the fraction of water gain (ml) relative to the cantaloupe weight in g (\eqn{default=0.004} according to \insertCite{Richards2004;textual}{qraLm}).}
}
\value{
A list of two elements:
\describe{
\item{N}{(\code{CFU}) A matrix of size \code{nLots} by \code{sizeLot} containing the numbers of \emph{L. monocytogenes} cells on cantaloupe from the irrigation water route;}
\item{P}{Prevalence of field lots of cantaloupes contaminated with \emph{L. monocytogenes} from the irrigation water route.}
}
}
\description{
The function \code{\link[=caIrrig2rind]{caIrrig2rind()}} evaluates the contamination of cantaloupes through irrigation water only.
It considers water contamination characteristics: prevalence \code{pIrrig} and concentration \code{cIrrig}.
\code{pIrrig} and \code{cIrrig} have to be chosen by the user according to existing data of prevalence. \code{cIrrig} is
conditional to water sources contaminated with \emph{L. monocytogenes}.
}
\note{
\insertCite{Background2022;textual}{qraLm} lists many estimates of \emph{L. monocytogenes} in water environments.
The estimate of \eqn{0.131} provided by \insertCite{Raschle2021;textual}{qraLm} has been chosen as default.
The distribution about the concentration of \emph{L. monocytogenes} in irrigation water is represented as a uniform distribution,
using the minimum and maximum values from \insertCite{Sharma2020;textual}{qraLm}, who reported \eqn{<0.03\ to\ 11} MPN \emph{L. monocytogenes}/ L water.
The algorithm assumes that the amount of water deposited on the cantaloupe rind after the last irrigation is as a percentage of the cantaloupe weight,
and is sampled from a uniform distribution of parameters between a minimum value of zero and a maximum value of \eqn{0.004},
taken from \insertCite{Richards2004;textual}{qraLm}.
}
\examples{
dat <- caPrimaryProduction(
  nLots = 100,
  sizeLot = 100,
  pSoil = 0.089,
  pManure = 0.1,
  pIrrigRaining = 0.05,
  pFoil = 0.5,
  rFoil = 0.75,
  pIrrig = 0.131,
  cantaWeight = 800
)

caIrrig2rind(
  nLots = 5,
  sizeLot = 10,
  pIrrig = 0.5,
  cIrrigLogMin = 1,
  cIrrigLogMax = 2
)

}
\references{
\insertRef{stats}{qraLm}

\insertRef{mc2d}{qraLm}

\insertRef{Background2022}{qraLm}

\insertRef{Raschle2021}{qraLm}

\insertRef{Richards2004}{qraLm}

\insertRef{Sharma2020}{qraLm}
}
\author{
Laurent Guillier
}
\keyword{cross-contamination}
\keyword{irrigation}
\keyword{transfer}
