% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caPrimaryProduction.R
\name{caPrimaryProduction}
\alias{caPrimaryProduction}
\title{Contamination of cantaloupe rind from soil and irrigation water}
\usage{
caPrimaryProduction(
  nLots,
  sizeLot,
  cantaWeight = 1000,
  pSoil = 0.089,
  pManure = 0.5,
  fManure = 7,
  cSoilLogMin = -1,
  cSoilLogMode = 0.6,
  cSoilLogMax = 1.48,
  qSoilMin = 0.05,
  qSoilMode = 0.5,
  qSoilMax = 5,
  pFoil = 0.5,
  rFoil = 0.9,
  fIrrigRaining = 25,
  pIrrigRaining = 0.1,
  pIrrig = 0.131,
  cIrrigLogMin = -1.52,
  cIrrigLogMax = 1.04,
  pWaterGainMin = 0,
  pWaterGainMax = 0.004
)
}
\arguments{
\item{nLots}{Size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of cantaloupes per cultivation lot or field (scalar).}

\item{cantaWeight}{(\code{g}) weight of a cantaloupe}

\item{pSoil}{Prevalence of contamination of soil (provided by the user, \code{default=0.089} according to \insertCite{Strawn2013a;textual}{qraLm}).}

\item{pManure}{Proportion of fields using organic amendments (provided by user, \code{default=0.5})}

\item{fManure}{Odds-ratio estimate associated to use or organic amendment in soil (\code{default=7.0} according to \insertCite{Strawn2013b;textual}{qraLm}).}

\item{cSoilLogMin}{(\verb{log10 CFU/g}) Minimum value of the triangular distribution describing variability of concentration
(according to \insertCite{Dowe1997;textual}{qraLm}: \code{default=-1} log10\ CFU/g).}

\item{cSoilLogMode}{(\verb{log10 CFU/g}) Mode value of the triangular distribution describing variability of concentration
(according to \insertCite{Dowe1997;textual}{qraLm}: \code{default=0.6} log10\ CFU/g).}

\item{cSoilLogMax}{(\verb{log10 CFU/g}) Maximum value of the triangular distribution describing variability of concentration
(according to  \insertCite{Dowe1997;textual}{qraLm}: \code{default=1.48} log10\ CFU/g).}

\item{qSoilMin}{(\code{g}) Minimum value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\code{default=0.05} g).}

\item{qSoilMode}{(\code{g}) Mode value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\code{default=0.5} g).}

\item{qSoilMax}{(\code{g}) Maximum value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\code{default=5} g).}

\item{pFoil}{Proportion of fields grown in foil (e.g. plastic mulch) (\code{default=0.5}).}

\item{rFoil}{Reduction fraction of the quantity of soil transferred to rind (\verb{default=default EKE: 0.9}).}

\item{fIrrigRaining}{Odds-ratio estimate associated to use of irrigation and raining events up to \code{2} days before harvest (\code{default=25.0} according
to \insertCite{Weller2015;textual}{qraLm}).}

\item{pIrrigRaining}{Proportion of fields that undergo irrigation or raining just previous harvest (provided by user, \code{default=0.1}).}

\item{pIrrig}{prevalence of contamination in irrigation water (provided by the user, \code{default=0.131} according to \insertCite{Raschle2021;textual}{qraLm})}

\item{cIrrigLogMin}{(\verb{log CFU/L}) min value of the uniform distribution (provided by the user, \code{default=-1.52} according to \insertCite{Sharma2020;textual}{qraLm})}

\item{cIrrigLogMax}{(\verb{log CFU/L})  max value of the uniform distribution (provided by the user, \code{default=1.04} according to \insertCite{Sharma2020;textual}{qraLm})}

\item{pWaterGainMin}{minimum value of the fraction of water gain (\code{ml}) relative to the cantaloupe weight in \code{g} (\code{default=0.0})}

\item{pWaterGainMax}{maximum value of the fraction of water gain (\code{ml}) relative to the cantaloupe weight in \code{g}
(\code{default=0.004} according to \insertCite{Richards2004;textual}{qraLm})}
}
\value{
A list of two elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} by \code{sizeLot} containing the numbers of \emph{L. monocytogenes} cells on cantaloupe at harvest;}
\item{\code{P}}{Prevalence of field lots of cantaloupes contaminated with \emph{L. monocytogenes} cells at harvest (scalar).}
}
}
\description{
\code{\link[=caPrimaryProduction]{caPrimaryProduction()}} is a function that assesses the contamination of cantaloupes through soil and irrigation water.
It considers soil and water contamination characteristics such as prevalence (\code{pSoil} and \code{pIrrig})
and distributions of \emph{L. monocytogenes} concentration in soil (\code{cSoilLogMin}, \code{cSoilLogMode} and \code{cSoilLogMax}
and water (\code{cIrrigLogMin} and \code{cIrrigLogMax}) as inputs.
\code{pSoil} and \code{pIrrig} have to be chosen by the user according to existing data of prevalence; whereas \code{pSoil} is conditional to several risk factors.
The identified risk factors (\insertCite{Background2022}{qraLm}) are:
\enumerate{
\item irrigation in the previous days before harvesting; and
\item the use of organic fertilizer
}
which affect \code{pSoil} as associated odds-ratios (\code{F_irrig_rain} and \code{fManure}). In this context, the user has to define the proportion of fields
practicing irrigation prior to harvest or undergoing rain (\code{p_irrig_rain}), and the proportion of fields using organic amendments (\code{pManure}).
Default values for \code{c_soil} and \code{c_irrig} have been proposed based on \insertCite{Background2022;textual}{qraLm}. \code{c_irrig} is
conditional to water sources contaminated with \emph{L. monocytogenes}.
}
\note{
The prevalence in soil, \code{pSoil}, must be provided by the user according to existing data. \insertCite{Background2022;textual}{qraLm} lists various prevalence
estimates in different regions of the world. A default value of 0.089 \insertCite{Strawn2013a}{qraLm} is taken.
\code{pSoil} is conditional to several risk factors; yet this functions is based upon two risk factors:
\enumerate{
\item irrigation or rain occurring prior to harvesting, whose odds-ratio has been estimated at \code{25.0} when taking place \code{24} hours before harvesting \insertCite{Weller2015}{qraLm}; and
\item application of organic fertilizer, whose odds-ratio has been estimated at \code{7.0} when manure is applied within \code{1} year \insertCite{Strawn2013b}{qraLm}.
}
The distribution about the concentration of \emph{L. monocytogenes} in soil is represented as a triangular distribution, using data from
\insertCite{Dowe1997;textual}{qraLm}, who reported a mean of \code{4.0} MPN \emph{L. monocytogenes} \ g soil, with a 95\% confidence interval of \verb{<1.0-28} MPN/g.
\insertCite{Background2022;textual}{qraLm} lists many estimates of \emph{L. monocytogenes} in water environments.
The estimate of \code{0.131} provided by \insertCite{Raschle2021;textual}{qraLm} has been chosen as default.
The distribution about the concentration of \emph{L. monocytogenes} in irrigation water is represented as a uniform distribution,
using the minimum and maximum values from \insertCite{Sharma2020;textual}{qraLm}, who reported \verb{<0.03} to \code{11} MPN \emph{L. monocytogenes} / L water.
The algorithm assumes that the amount of water deposited on the cantaloupe rind after the last irrigation is as a percentage of the cantaloupe weight,
and is sampled from a uniform distribution of parameters between a minimum value of zero and a maximum value of \code{0.004}, taken from \insertCite{Richards2004;textual}{qraLm}.
}
\examples{
dat <- caPrimaryProduction(
  nLots = 100,
  sizeLot = 100,
  cantaWeight = 800,
  pSoil = 0.089,
  pManure = 0.1,
  pIrrigRaining = 0.05,
  pFoil = 0.5,
  rFoil = 0.75,
  pIrrig = 0.131
)
hist(dat$N)

}
\references{
\insertRef{stats}{qraLm}
\insertRef{mc2d}{qraLm}
\insertRef{Background2022}{qraLm}
\insertRef{Dowe1997}{qraLm}
\insertRef{Raschle2021}{qraLm}
\insertRef{Richards2004}{qraLm}
\insertRef{Sharma2020}{qraLm}
\insertRef{Strawn2013a}{qraLm}
\insertRef{Strawn2013b}{qraLm}
\insertRef{Weller2015}{qraLm}
}
\seealso{
Other data generation: 
\code{\link{Lot2LotGen}()},
\code{\link{LotGen}()}
}
\author{
Laurent Guillier
}
\concept{data generation}
\keyword{cross-contamination}
\keyword{irrigation,}
\keyword{soil,}
\keyword{splash,}
\keyword{transfer,}
