% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caSoil2rind.R
\name{caSoil2rind}
\alias{caSoil2rind}
\title{Route of contamination of cantaloupe from soil to rind}
\usage{
caSoil2rind(
  nLots,
  sizeLot,
  pSoil = 0.089,
  fManure = 7,
  pManure = 0.5,
  fIrrigRaining = 25,
  pIrrigRaining = 0.1,
  cSoilLogMin = -1,
  cSoilLogMode = 0.6,
  cSoilLogMax = 1.48,
  qSoilMin = 0.05,
  qSoilMode = 0.5,
  qSoilMax = 5,
  pFoil = 0.5,
  rFoil = 0.9
)
}
\arguments{
\item{nLots}{Size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of cantaloupes per cultivation lot or field (scalar).}

\item{pSoil}{Probability of contamination of soil (provided by the user, \eqn{default=0.089} according to \insertCite{Strawn2013a;textual}{qraLm})}

\item{fManure}{Odds-ratio estimate associated to use or organic amendment in soil (\eqn{default=7.0} according to \insertCite{Strawn2013b;textual}{qraLm})}

\item{pManure}{Proportion of fields using organic amendments (provided by user, \eqn{default=0.5})}

\item{fIrrigRaining}{Odds-ratio estimate associated to use of irrigation and raining events up to 2 days before harvest
(\eqn{default=25.0} according to \insertCite{Weller2015;textual}{qraLm}).}

\item{pIrrigRaining}{Proportion of fields that undergo irrigation or raining just previous harvest (provided by user, \eqn{default=0.1}).}

\item{cSoilLogMin}{(log10 CFU/g) Minimum value of the triangular distribution describing variability of concentration
(according to \insertCite{Dowe1997;textual}{qraLm} \eqn{default=-1\ log10\ CFU/g}).}

\item{cSoilLogMode}{(log10 CFU/g) Mode value of the triangular distribution describing variability of concentration
(according to \insertCite{Dowe1997;textual}{qraLm} \eqn{default=0.6\ log10\ CFU/g}).}

\item{cSoilLogMax}{(log10 CFU/g) Maximum value of the triangular distribution describing variability of concentration
(according to  \insertCite{Dowe1997;textual}{qraLm}: \eqn{default=1.48\ log10\ CFU/g}).}

\item{qSoilMin}{(\code{g}) Minimum value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\eqn{default=0.05\ g}).}

\item{qSoilMode}{(\code{g}) Mode value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\eqn{default=0.5\ g}).}

\item{qSoilMax}{(\code{g}) Maximum value of the triangular distribution describing variability of quantity of soil deposited on cantaloupe (\eqn{default=5\ g}).}

\item{pFoil}{Proportion of fields grown in foil (e.g. plastic mulch) (EKE: \eqn{default=0.5})}

\item{rFoil}{Reduction fraction of the quantity of soil transferred to rind (EKE: \eqn{default=0.9})}
}
\value{
A list of two elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} by \code{sizeLot} containing the numbers of \emph{L. monocytogenes} cells on cantaloupe
from the soil route;}
\item{\code{P}}{Prevalence of field lots of cantaloupes contaminated with \emph{L. monocytogenes} from the soil route (scalar).}
}
}
\description{
The function \code{\link[=caSoil2rind]{caSoil2rind()}} evaluates the contamination of cantaloupes through soil only.
It considers soil contamination characteristics such as prevalence \code{pSoil} and concentration \code{c_soil} as well as the
quantity of soil deposited on the cantaloupe rind. Furthermore, the algorithm allows for the use of foils to protect
cantaloupe from soil contamination, and the user must provide the probability of growing cantaloupes using such protective foils.
\code{pSoil} and \code{c_soil} have to be chosen by the user according to existing data; whereas \code{pSoil} is conditional to several risk factors.
The identified risk factors \insertCite{Background2022}{qraLm} are:
\enumerate{
\item irrigation in the previous days before harvesting; and
\item the use of organic fertilizer.
}
which affect \code{pSoil} as associated odds-ratios (\code{F_irrig_rain} and \code{fManure}). In this context, the user has to define the proportion of fields
practising irrigation prior to harvest or undergoing rain (\code{p_irrig_rain}), and the proportion of fields using organic amendments (\code{pManure}).
}
\note{
The prevalence in soil, \code{pSoil}, must be provided by the user according to existing data. \insertCite{Background2022;textual}{qraLm}
lists various prevalence estimates in different regions of the world. A \code{default=0.089} \insertCite{Strawn2013a}{qraLm} is taken.
\code{pSoil} is conditional to several risk factors; yet this functions is based upon two risk factors:
\enumerate{
\item irrigation or rain occurring prior to harvesting, whose odds-ratio has been estimated at 25.0 when taking place 24 hours before
harvesting \insertCite{Weller2015}{qraLm}; and
\item application of organic fertilizer, whose odds-ratio has been estimated at 7.0 when manure is applied within 1 year \insertCite{Strawn2013b}{qraLm}.
}
The distribution about the concentration of \emph{L. monocytogenes} in soil is represented as a triangular distribution,
using data from \insertCite{Dowe1997;textual}{qraLm}, who reported a mean of 4.0 MPN \emph{L. monocytogenes}/ g soil, with a 95\%
confidence interval of \eqn{<1.0-28\ MPN/g}.
}
\examples{
# Considering all fields with foil: less cultivation lots contaminated than pSoil
caSoil2rind(nLots = 20, sizeLot = 10, pSoil = 0.089, pManure = 0.0, pIrrigRaining = 0, pFoil = 1.0)

# Effect of manure (risk factor): more batches contaminated than pSoil
caSoil2rind(nLots = 20, sizeLot = 10, pSoil = 0.089, pManure = 0.5, pIrrigRaining = 0, pFoil = 0.0)

}
\references{
\insertRef{stats}{qraLm}

\insertRef{mc2d}{qraLm}

\insertRef{Background2022}{qraLm}

\insertRef{Dowe1997}{qraLm}

\insertRef{Strawn2013a}{qraLm}

\insertRef{Strawn2013b}{qraLm}

\insertRef{Weller2015}{qraLm}
}
\author{
Laurent Guillier
}
\keyword{cross-contamination}
\keyword{soil}
\keyword{transfer}
