% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caTrans2RetRTE.R
\name{caTrans2RetRTE}
\alias{caTrans2RetRTE}
\title{Transport of RTE diced cantaloupe packs from the end of processing to retail}
\usage{
caTrans2RetRTE(
  data = list(),
  MPD,
  unitSize = NULL,
  meanEGR5,
  seEGR5,
  Tmin,
  tempMin,
  tempMode,
  tempMax,
  timeMin,
  timeMode,
  timeMax,
  lnQ0Mean,
  lnQ0Sd
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{N}{(\code{CFU}) A matrix containing the numbers of \emph{L. monocytogenes} in packs of RTE diced cantaloupe before transport,
from contaminated lots.}
\item{P}{Prevalence of contaminated lots (scalar).}
}}

\item{MPD}{(\eqn{log10\ CFU/g}) Maximum population density of \emph{L. monocytogenes} in cantaloupe flesh (scalar).}

\item{unitSize}{(\code{g}) Weight of a pack of cantaloupe dices (scalar).}

\item{meanEGR5}{(\eqn{h^{-1}}) Mean exponential growth rate of \emph{L. monocytogenes} in cantaloupe flesh at
5 \eqn{^\circ C} (suggested \eqn{default=0.03557288\ h^{-1}}) (scalar).}

\item{seEGR5}{(\eqn{h^{-1}}) Standard error about the mean of the exponential growth rate of \emph{L. monocytogenes} in
cantaloupe flesh at 5 \eqn{^\circ C} (suggested \eqn{default=0.004\ h^{-1}}) (scalar).}

\item{Tmin}{(\eqn{^\circ C}) Nominal minimum temperature for growth of \emph{L. monocytogenes} in cantaloupe flesh
(suggested \eqn{default=-2.0196\ ^\circ C}) (scalar).}

\item{tempMin}{(\eqn{^\circ C}) Minimum transportation temperature (scalar).}

\item{tempMode}{(\eqn{^\circ C}) Mode of the transportation temperature (scalar).}

\item{tempMax}{(\eqn{^\circ C}) Maximum transportation temperature (scalar).}

\item{timeMin}{(\code{h}) Minimum transportation time (scalar).}

\item{timeMode}{(\code{h}) Mode of the transportation time (scalar).}

\item{timeMax}{(\code{h}) Maximum transportation time (scalar).}

\item{lnQ0Mean}{Mean of the natural log of \code{Q0} (suggested \eqn{default=-0.096728} at \eqn{time=0}) (scalar).}

\item{lnQ0Sd}{Standard deviation of the natural log of \code{Q0} (suggested \eqn{default=0.063930} at \eqn{time=0}) (scalar).}
}
\value{
A list of five elements:
\describe{
\item{N}{(\code{CFU}) A matrix containing the numbers of \emph{L. monocytogenes} in packs of RTE diced cantaloupe after transport,
from contaminated lots;}
\item{P}{Prevalence of RTE diced cantaloupe lots contaminated with \emph{L. monocytogenes} (scalar);}
\item{lnQt}{Natural log of the \code{Q} parameter at the end of transportation (matrix);}
\item{lotEGR5}{(\eqn{h^{-1}}) Growth rate of \emph{L. monocytogenes} in cantaloupe flesh specific to every lot (vector);}
\item{unitSize}{(\code{g}) Weight of a pack of cantaloupe dices.}
}
}
\description{
The function \code{\link[=caTrans2RetRTE]{caTrans2RetRTE()}} simulates the growth of \emph{L. monocytogenes} in RTE diced cantaloupe during cold transport
to retail, and is based on the function \code{\link[=caGrowthBaranyi]{caGrowthBaranyi()}}. The algorithm considers that RTE diced cantaloupe packs
from every lot are subjected
to the same initial \code{q0} (related to lag phase), the same transportation temperature (\code{Temp}), and the same transportation \code{time}.
The normal distribution with parameters \code{lnQ0Mean} and \code{lnQ0Sd} represents the variability about \code{ln_q0}, a parameter
related to the physiological state of cells in the Baranyi ad Roberts' growth model. The uncertainty about \code{EGR5} is
represented by a normal distribution with parameters \code{meanEGR5} and \code{sdEGR5}. Pert distributions represent the
lot-specific variability in transport time and temperature.
}
\note{
\code{meanEGR5}, \code{seEGR5} and \code{Tmin} were determined from fitting a square-root model to data extracted from multiple sources, as described in
more detail in the function \code{\link[=caGrowthBaranyi]{caGrowthBaranyi()}}. \code{lnQ0Mean} and \code{lnQ0Sd} were determined by running a Monte Carlo simulation to data in canary melon
consisting of growth rate and lag phase duration obtained at different temperatures. Such data were extracted from \insertCite{Scolforo2017;textual}{qraLm}.
The suggested parameters for the Pert distribution of transportation temperature are taken from \insertCite{Ding2013;textual}{qraLm} and \insertCite{SantAna2014;textual}{qraLm}:
\eqn{Temp\_min=3\ ^\circ C}, \eqn{Temp\_mode=5\ ^\circ C} and \eqn{Temp\_max=10.3\ ^\circ C}. Parameters for the transportation time distribution should be defined by the user.
In addition to the final \emph{L. monocytogenes} numbers, the function also returns the lot-specific values of \code{lotEGR5} and \code{lnQt} so that the
\emph{L. monocytogenes} growth could be followed up in subsequent logistics stages.
}
\examples{
timeMin <- 2
timeMode <- 5
timeMax <- 9
nLots <- 1000
sizeLot <- 250
dat <- list(
  N = matrix(rpois(nLots * sizeLot, 23), nLots, sizeLot),
  P = 0.4,
  unitSize = 200
)
ArriveRetail <- caTrans2RetRTE(dat,
  MPD = 8.5,
  meanEGR5 = 0.03557288,
  seEGR5 = 0.004,
  Tmin = -2.0196,
  tempMin = 3,
  tempMode = 5,
  tempMax = 10.3,
  timeMin = timeMin,
  timeMode = timeMode,
  timeMax = timeMax,
  lnQ0Mean = -0.096728,
  lnQ0Sd = 0.063930
)
str(ArriveRetail)
hist(ArriveRetail$N)
}
\references{
\insertRef{Ding2013}{qraLm}

\insertRef{extraDistr}{qraLm}

\insertRef{mc2d}{qraLm}

\insertRef{SantAna2014}{qraLm}

\insertRef{Scolforo2017}{qraLm}

\insertRef{stats}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Baranyi}
\keyword{growth}
\keyword{lag-phase}
\keyword{logistics}
\keyword{stochastic}
