% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRForModel.R
\name{DRForModel}
\alias{DRForModel}
\title{Dose-response model function for the qraLm models}
\usage{
DRForModel(data = list(), model = "JEMRA", population = 1, Poisson = FALSE)
}
\arguments{
\item{data}{a list of (a minima):
\describe{
\item{\code{N}}{(\code{CFU}) A matrix containing the numbers of \emph{L. monocytogenes} per portion of defrosted or frozen vegetables.}
}}

\item{model}{either \code{JEMRA}, \code{Pouillot}, \code{Fritsch}, \code{EFSA}, \code{EFSAMV},\code{EFSAV} or \code{EFSALV}}

\item{population}{considered population (scalar).}

\item{Poisson}{if \code{TRUE}, assume that \code{Dose} is the mean of a Poisson distribution.
(actual LogNormal Poisson). If \code{FALSE} (default), assume that \code{Dose} is the actual number of bacteria.}
}
\value{
the data object with added:
\describe{
\item{\code{Risk}}{A matrix of risk, of size similar to \code{N}}
\item{\code{Model}}{the Model}
\item{\code{Population}}{the Population}
}
}
\description{
This function provides the marginal probability of invasive listeriosis
in a given \code{population} for a given \code{Dose} in \code{CFU} using the
\code{JEMRA}, the \code{Pouillot}, the \code{Fritsch} or the \code{EFSA} dose-response models
or the model developed within this project (\code{EFSAMV},\code{EFSAV},\code{EFSALV}) (see References).
}
\details{
see \code{\link[doseresponsemodels:DR]{doseresponsemodels::DR()}} or \code{\link[doseresponsemodels:DRQuick]{doseresponsemodels::DRQuick()}} for details on population and models.
}
\note{
This function uses (for all model but \code{JEMRA}) a linear approximation (\code{approxfun})
from the exact \code{\link[doseresponsemodels:DR]{doseresponsemodels::DR()}} model evaluated on \eqn{Dose = c(0,10^{seq(-5,12,length=1701)})}
(if \code{Poisson=TRUE}) or \eqn{c(0,10^{seq(0,12,length=2000)})} (if \code{Poisson=FALSE}).
Any Dose lower or higher than these ranges will lead to \code{NA}.
}
\examples{
data <- list(N = matrix(10^runif(100, 0, 5), ncol = 25))
DRForModel(data, "JEMRA", 1)
DRForModel(data, "EFSAMV", 10)

}
\references{
\insertRef{EFSA2018}{qraLm}
\insertRef{FAO-WHO2004}{qraLm}
\insertRef{Fritsch2018}{qraLm}
\insertRef{Pouillot2015}{qraLm}
}
\author{
Regis Pouillot, Vasco Cadavez
}
\keyword{dose-response}
\keyword{model}
