% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expandedjamesonC}
\alias{expandedjamesonC}
\title{Integrate the growth model with Jameson effect}
\usage{
expandedjamesonC(init, duration, step_size = 1, start = 0)
}
\arguments{
\item{init}{initial values}

\item{duration}{the end of the integration}

\item{step_size}{the step size of the integration}

\item{start}{start of the integration}
}
\description{
This function integrate the Jameson model through ODE. See \link{sfGrowthJameson}.
}
