% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expandedjamesonC_set_params}
\alias{expandedjamesonC_set_params}
\title{Set the parameters to integrate the growth model with Jameson effect}
\usage{
expandedjamesonC_set_params(
  mumax1 = 0.14,
  mumax2 = 0.3,
  ymax1 = 1e+08,
  ymax2 = 1e+07,
  gamma = 1
)
}
\arguments{
\item{mumax1}{mumax (/h) of the first population}

\item{mumax2}{mumax (/h) of the second population}

\item{ymax1}{MPD (CFU) of the first population}

\item{ymax2}{MPD (CFU) of the second population}

\item{gamma}{interaction}
}
\description{
This function set the parameters to integrate the Jameson model through ODE. See \link{sfGrowthJameson}.
}
