% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvBlanching.R
\name{fvBlanching}
\alias{fvBlanching}
\title{Effect of blanching on \emph{L. monocytogenes} in processed vegetables}
\usage{
fvBlanching(
  data = list(),
  nLots = NULL,
  tempBlanch = 83,
  timeBlanch = 0.75,
  logDrefMean = -1.78,
  logDrefSd = 0.251,
  zT = 6.06
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units containing the numbers  of
\emph{L. monocytogenes} in pre-blanching units or portions from contaminated lots;}
\item{\code{ProbUnitPos}}{Probability that the lot is contaminated (a lot is considered contaminated
if at least one unit or portion is contaminated) (vector);}
\item{\code{P}}{Mean prevalence of contaminated lots (scalar).}
}}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{tempBlanch}{(\eqn{^\circ C}) Temperature of blanching (scalar or vector).}

\item{timeBlanch}{(\code{min}) Duration of blanching (scalar or vector).}

\item{logDrefMean}{Mean parameter of the normal distribution representing the variability in the log 10 of the
reference \code{D} (\code{min}) at 70 \eqn{^\circ C} (scalar or vector).}

\item{logDrefSd}{Standard deviation of the normal distribution representing the variability in
the log 10 of the reference \code{D} (min) at 70 \eqn{^\circ C} (scalar or vector).}

\item{zT}{fixed \code{z} value characterising the effect of temperature on \code{D} for \emph{L. monocytogenes} in vegetables (scalar).}
}
\value{
A list of four elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes}
per unit or portion after blanching;}
\item{\code{ProbUnitPos}}{Probability that the lot is contaminated after blanching (a lot is considered contaminated if at least
one unit or portion remained contaminated after blanching) (vector);}
\item{\code{P}}{Mean prevalence of contaminated lots after blanching (scalar);}
\item{\code{pSurviveBlanching}}{Probability of a microbial cell to survive the blanching process (vector).}
}
}
\description{
The function \code{\link[=fvBlanching]{fvBlanching()}} describes the effect of blanching on the numbers of \emph{L. monocytogenes}
present in vegetables such as broccoli, mushroom, onions, peas and pepper. The function is based on the Bigelow
model, which describes the decimal reduction time (\code{D}) as a function of temperature, with parameters \code{z} and reference \code{D} (\code{Dref})
at 70 \eqn{^\circ C}.
This function assumes that the extent of reduction is global for the aforementioned vegetables, and depends on the duration and
temperature of blanching. Whereas the algorithm considers the \code{z} value as fixed, \code{Dref} is assumed to be strain-specific,
and thus different between lots.
}
\note{
If no more bacteria are in a lot, one cell will be set in one unit of that lot. The heat resistance parameters suggested for use
(\eqn{zT=6.06\ ^{\circ} C}, and \eqn{log10\ Dref\_mean=-1.78} and \eqn{log10Dref\_sd=0.251} at the reference temperature of 70 \eqn{^{\circ} C})
were obtained from fitting a Bigelow model to survival data of \emph{L. monocytogenes} in broccoli, mushrooms, onions, peas and pepper
extracted from \insertCite{Mazzotta2001;textual}{qraLm}
}
\examples{
dat <- Lot2LotGen(
  nLots = 1000, sizeLot = 1000, unitSize = 500,
  betaAlpha = 0.5112, betaBeta = 9.959, C0MeanLog = 1.023, C0SdLog = 0.3267,
  propVarInter = .7
)
res <- fvBlanching(dat,
  tempBlanch = 85,
  timeBlanch = 1,
  logDrefMean = -1.78,
  logDrefSd = 0.251,
  zT = 6.06
)
res$P
hist(c(res$N))

}
\references{
\insertRef{stats}{qraLm}

\insertRef{Mazzotta2001}{qraLm}
}
\author{
Laurent Guillier
}
\keyword{blanching}
\keyword{cooling}
\keyword{heat}
\keyword{treatment}
