% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvDefrost.R
\name{fvDefrost}
\alias{fvDefrost}
\title{Defrosting of frozen vegetables at home}
\usage{
fvDefrost(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  Temp,
  time,
  MPD,
  Tmin,
  meanEGR5,
  sdEGR5,
  servingSize,
  pDefrost
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{\code{N}}{(\code{CFU})  A matrix of size \code{nLots} lots by \code{sizeLot} units containing the numbers of \emph{L. monocytogenes} per
portion of frozen vegetables to be defrosted;}
\item{\code{ProbUnitPos}}{Lot-specific probability of contaminated portions or servings before defrosting (vector);}
\item{\code{P}}{Mean prevalence of contaminated portions or servings (scalar).}
}}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{Temp}{(\eqn{^\circ C}) temperature at which frozen vegetables are defrosted (scalar or vector).}

\item{time}{(\code{h}) time that a portion of frozen vegetables is kept for defrosting (scalar or vector).}

\item{MPD}{(\verb{log10 CFU/g}) is the  maximum population density of \emph{L. monocytogenes} in blanched vegetables (scalar or vector).}

\item{Tmin}{(\eqn{^\circ C}) minimum temperature for growth of \emph{L. monocytogenes} in blanched vegetables at
5 \eqn{^\circ C} (scalar or vector).}

\item{meanEGR5}{(\eqn{log10\ CFU/g/h}) mean of the exponential growth rate of \emph{L. monocytogenes} in blanched vegetables at
5 \eqn{^\circ C} (scalar or vector).}

\item{sdEGR5}{(\eqn{log10\ CFU/g/h}) standard deviation of the exponential growth rate of \emph{L. monocytogenes} in
blanched vegetables at 5 \eqn{^\circ C} (scalar or vector).}

\item{servingSize}{(\code{g}) weight of frozen vegetables portioned from the pack or serving size (scalar or vector).}

\item{pDefrost}{probability of defrosting frozen vegetables (scalar).}
}
\value{
A list of three elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes}
in the portions of defrosted or non-defrosted vegetables.}
\item{\code{ProbUnitPos}}{Lot-specific probability of contaminated portions or servings, defrosted or not (vector).}
\item{\code{P}}{Mean prevalence of contaminated portions of defrosted or non-defrosted vegetables (scalar).}
}
}
\description{
The \code{\link[=fvDefrost]{fvDefrost()}} function simulates the possibility that consumers thaw frozen vegetables prior to cooking or consumption. If consumers defrost
frozen vegetables, the algorithm estimates the growth of \emph{L. monocytogenes} during the time frozen vegetables are held for thawing.
The function employs the lognormal distribution to represent the variability (at the portion level) in exponential growth rate at 5 \eqn{^\circ C\ log10\ CFU/g/h}
of \emph{L. monocytogenes} in heat-treated vegetables including corn, green peas, carrots, broccoli, beans and asparagus stored in air,
as determined in \insertCite{EFSA2020;textual}{qraLm}. A log-linear growth model is employed, without any lag phase duration.
}
\note{
Combining literature data, \insertCite{EFSA2020;textual}{qraLm} determined that the variability in the exponential growth rate
of \emph{L. monocytogenes} in heat-treated vegetables at 5 \eqn{^\circ C} can be assumed to distribute as a lognormal distribution
with mean ``meanEGR5=0.0117\verb{log10\\ CFU/g/h and standard deviation}sdEGR5=0.00816` log10\ CFU/g/h.
Furthermore, \insertCite{EFSA2020;textual}{qraLm} assumed a minimum temperature for growth `Tmin=-1.18` \eqn{^\circ C}.
These are the values recommended for use as arguments of the function.
}
\examples{
Tmin <- -1.18
meanEGR5 <- 0.0117
sdEGR5 <- 0.00816
dat <- Lot2LotGen(
                  nLots = 50,
                  sizeLot = 100,
                  unitSize = 500,
                  betaAlpha = 0.5112,
                  betaBeta = 9.959,
                  C0MeanLog = 1.023,
                  C0SdLog = 0.3267,
                  propVarInter = 0.7
                  )

defrosted <- fvDefrost(
  dat,
  Temp = 15,
  time = 2,
  MPD = 8,
  Tmin = -1.18,
  meanEGR5 = meanEGR5,
  sdEGR5 = sdEGR5,
  servingSize = 100,
  pDefrost = 0.8
)
hist(defrosted$N) # histogram of microbial cells in contaminated defrosted/non-defrosted servings

}
\references{
\insertRef{EFSA2020}{qraLm}

\insertRef{stats}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{counter}
\keyword{defrost}
\keyword{thawing}
\keyword{top}
\keyword{unfreezing}
