% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvPartitioningCC.R
\name{fvPartitioningCC}
\alias{fvPartitioningCC}
\title{Cross-contamination during freezing/packaging and partitioning into packed units}
\usage{
fvPartitioningCC(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  probCC,
  trMean = -0.44,
  trSd = 0.4,
  nEquip,
  bCCFV
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units containing the numbers of \emph{L. monocytogenes} in blanched
units or portions from contaminated lots;}
\item{\code{P}}{Mean prevalence of contaminated lots (scalar);}
\item{\code{ProbUnitPos}}{Probability of individual lots being contaminated (vector).}
}}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{probCC}{Probability of cross-contamination of vegetables from conveyors, freezer or packaging machine  (scalar).}

\item{trMean}{Mean parameter of the normal distribution representing the variability in the log 10 of the transfer coefficient of \emph{L. monocytogenes}
cells from surfaces to (frozen) vegetables  (scalar or vector).}

\item{trSd}{Standard deviation parameter of the normal distribution representing the variability in the log 10 of the transfer coefficient of
\emph{L. monocytogenes} from surfaces to (frozen) vegetables (scalar or vector).}

\item{nEquip}{(\code{CFU}) Numbers of \emph{L. monocytogenes} cells on the surface of conveyors, freezer or packaging machine in contact
with the (frozen) vegetables  (scalar or vector).}

\item{bCCFV}{Dispersion factor of the beta distribution representing the degree of heterogeneity in the number of cells between pack units (scalar).}
}
\value{
A list of three elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes} per pack unit,
from contaminated lots;}
\item{\code{ProbUnitPos}}{Probability of individual lots being contaminated after packaging (a lot is considered contaminated if at
least one pack unit is contaminated) (vector);}
\item{\code{P}}{Mean prevalence of contaminated lots after packaging (scalar).}
}
}
\description{
The \code{\link[=fvPartitioningCC]{fvPartitioningCC()}} function simulates the potential re-contamination of vegetables post-blanching, when in direct
contact with contaminated surfaces of conveyors, freezer or packaging machine, followed by the partitioning of the bulk
of frozen vegetables into packed units.
The cross-contamination algorithm accounts for four possible scenarios:
\enumerate{
\item cross-contamination occurring in lots already contaminated;
\item re-contamination occurring in lots that were not contaminated;
\item no cross-contamination occurring in lots already contaminated; and
\item no cross-contamination occurring in lots that were not contaminated.
}
Probabilities of occurrence of every event are computed. The partitioning algorithm randomly distributes the total numbers of cells from a
contaminated lot into packed units. The dispersion factor \code{b}, which is a parameter of the beta distribution, indicates the
extent of cell clustering in the bulk of frozen vegetables in a lot, and ultimately the heterogeneity in the number of cells
distributed among pack units.
}
\note{
The value of \eqn{probCC=0.125} is taken from \insertCite{Truchado2021;textual}{qraLm}, who analysed the presence of \emph{L. monocytogenes} on
conveyor belts, freezing tunnels and packaging machines in frozen food processing plants. The value of beta=1 represents moderate
clustering of cells in the bulk of frozen vegetables from a lot \insertCite{Nauta2005;textual}{qraLm}. \insertCite{Hoelzer2012;textual}{qraLm}
established the log 10 of the transfer coefficient of \emph{L. monocytogenes} from stainless steel to vegetables as a normal
distribution with \eqn{TR\_mean=-0.44} and \eqn{TR\_sd=0.40}.
}
\examples{
dat <- Lot2LotGen(
  nLots = 50,
  sizeLot = 100,
  unitSize = 500,
  betaAlpha = 0.5112,
  betaBeta = 9.959,
  C0MeanLog = 1.023,
  C0SdLog = 0.3267,
  propVarInter = 0.7
)
Nf <- fvPartitioningCC(dat,
                       probCC = 0.125, trMean = -0.44,
                       trSd = 0.40, nEquip = 500, bCCFV = 1
)
hist(Nf$N) # displays histogram of cells in a pack from contaminated lots
}
\references{
\insertRef{Truchado2021}{qraLm}
\insertRef{Hoelzer2012}{qraLm}
\insertRef{Nauta2005}{qraLm}
\insertRef{extraDistr}{qraLm}
\insertRef{mc2d}{qraLm}
\insertRef{Rdpack}{qraLm}
\insertRef{stats}{qraLm}
\insertRef{iRisk2021}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt} and Regis Pouillot \email{rpouillot.work@gmail.com}
}
\keyword{cross-contamination}
\keyword{re-contamination}
\keyword{transfer}
