% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvPortioning.R
\name{fvPortioning}
\alias{fvPortioning}
\title{Portioning of the contents of a pack at home}
\usage{
fvPortioning(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  servingSize,
  unitSize = NULL,
  bPort
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes} per
pack before handling at home;}
\item{\code{ProbUnitPos}}{Probability of (tested) individual lots being contaminated (vector);}
\item{\code{P}}{Mean prevalence of contaminated lots (scalar).}
}}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{servingSize}{(\code{g}) is the portion taken from a pack, which will later equals to the serving size (scalar or vector).}

\item{unitSize}{(\code{g}) is the weight of a pack of frozen vegetables (scalar).}

\item{bPort}{dispersion factor of cells within the package (scalar or vector).}
}
\value{
A list of three elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes}
in the portions of frozen vegetables.}
\item{\code{ProbUnitPos}}{Lot-specific probability of contaminated portions or servings (vector).}
\item{\code{P}}{Mean prevalence of contaminated portions or servings (scalar).}
}
}
\description{
The \code{\link[=fvPortioning]{fvPortioning()}} function represents the portioning of a pack of frozen vegetables into a smaller unit. It is assumed that the microbial
cells present in a contaminated pack are distributed into servings following a beta-binomial distribution, although the algorithm only retains
one portion per pack (and not all the portions that can be obtained from a pack). The dispersion factor \code{bPort} represents the extent of cell
clustering in the frozen vegetables within the package. The fact that the algorithm takes into account that a contaminated pack can produce a
non-contaminated portion, enables the estimation of the prevalence of contaminated portions. It is assumed that the dispersion parameter and
the number of portions (that can be obtained from a pack) are independent of the microbial numbers.
}
\note{
A dispersion factor \eqn{bPort=1} represents moderate clustering of cells \insertCite{Nauta2005;textual}{qraLm} in the frozen vegetables
within the package.
}
\examples{
dat <- Lot2LotGen(
  nLots = 50,
  sizeLot = 100,
  unitSize = 500,
  betaAlpha = 0.5112,
  betaBeta = 9.959,
  C0MeanLog = 1.023,
  C0SdLog = 0.3267,
  propVarInter = 0.7
)
Nf <- fvPortioning(dat, servingSize = 150, unitSize = 500, bPort = 1)
str(Nf)
hist(Nf$N) # histogram of microbial cells in contaminated servings

}
\references{
\insertRef{Nauta2005}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt} and Regis Pouillot \email{rpouillot.work@gmail.com}
}
\keyword{handling}
\keyword{portions}
\keyword{servings}
