% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvReductionPostpack.R
\name{fvReductionPostpack}
\alias{fvReductionPostpack}
\title{Effect of a potential intervention step post-packaging}
\usage{
fvReductionPostpack(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  logIntervMin,
  logIntervMode,
  logIntervMax
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{N}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes} per pack unit, before the intervention;}
\item{ProbUnitPos}{Probability of individual lots being contaminated before the intervention (vector);}
\item{P}{Mean prevalence of contaminated lots before the intervention (scalar).}
}}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{logIntervMin}{(log10) minimum value of the triangular distribution representing log10 reduction (scalar).}

\item{logIntervMode}{(log10) mode value of the triangular distribution representing log10 reduction (scalar).}

\item{logIntervMax}{(log10) maximum value of the triangular distribution representing log10 reduction (scalar).}
}
\value{
A list of four elements:
\describe{
\item{N}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes}
per pack unit after post-packaging intervention;}
\item{ProbUnitPos}{Probability that the lot is contaminated after post-packaging intervention (a lot is considered contaminated if at least
one pack remained contaminated after the intervention) (vector);}
\item{P}{Mean prevalence of contaminated lots after post-packaging intervention (scalar);}
\item{pSurvReduction}{Probability of a microbial cell to survive the post-packaging intervention (vector).}
}
}
\description{
The function \code{\link[=fvReductionPostpack]{fvReductionPostpack()}} describes the effect of any potential intervention step that would
be applied on packed frozen vegetables in order to reduce the occurrence of \emph{L. monocytogenes}.
It is intended that the user proposes as inputs the minimum, mode and maximum of the log10 reduction for this intervention (MRA46).
}
\examples{
library(mc2d)
dat <- list(
  N = matrix(10^rnorm(1000 * 500, 1, 0), ncol = 1000, nrow = 500), P = 0.10,
  ProbUnitPos = rep(0.5, 500)
)
out <- fvReductionPostpack(dat,
  logIntervMin = 0,
  logIntervMode = 1,
  logIntervMax = 3
)
hist(out$N)
}
\references{
\insertRef{stats}{qraLm}

\insertRef{mc2d}{qraLm}
}
\author{
Laurent Guillier
}
\keyword{inactivation}
\keyword{reduction}
