% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvTesting.R, R/sfTesting.R
\name{fvTesting}
\alias{fvTesting}
\alias{sfTesting}
\title{Application of a testing regime for \emph{L. monocytogenes} in a sample of food units from a lot}
\usage{
fvTesting(
  data = list(),
  nTested = 5,
  gTested = 25,
  MTested = 0,
  cTested = 0,
  pLotTested,
  nLots = NULL,
  sizeLot = NULL,
  unitSize = NULL,
  Se = 1,
  gTestedEnum = 10,
  iterSub = NULL
)

sfTesting(...)
}
\arguments{
\item{data}{a list of:
\describe{
\item{N}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes} per pack unit, from contaminated lots.}
\item{ProbUnitPos}{Probability of individual lots being contaminated (a lot is considered contaminated if at least one pack unit is contaminated) (vector).}
\item{P}{Mean prevalence of contaminated lots (scalar).}
}}

\item{nTested}{sample size or number of units tested (scalar).}

\item{gTested}{(\code{g}) sample weight tested per unit (scalar or vector).}

\item{MTested}{(\code{CFU/g}) maximum concentration accepted in a sample (scalar or vector). If \eqn{>0}, will lead to a 3-class plan.}

\item{cTested}{maximum number of samples accepted between \eqn{m=0} and \code{M} (scalar or vector).}

\item{pLotTested}{proportion of lots subjected to sampling and testing or frequency of testing (scalar or vector).}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{unitSize}{(\code{g}) weight of the contents of a pack unit (scalar or vector).}

\item{Se}{Sensibility of the test or probability to detect each bacterial cell applying a given microbiological essay (scalar or vector).}

\item{gTestedEnum}{(\code{g}) sample weight tested for enumeration (scalar or vector).}

\item{iterSub}{number of internal repetition for the testing. If \code{NULL}, it will be set to the minimum of \code{nLots} and 1000.}

\item{...}{
  Arguments passed on to \code{\link[=fvTesting]{fvTesting}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A list of five elements of the data objects:
\describe{
\item{N}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the numbers of \emph{L. monocytogenes} per
pack unit, after within-lot testing.}
\item{P}{Mean prevalence of contaminated lots after within-lot testing (scalar).}
\item{ProbUnitPos}{Probability that the individual lot is contaminated given that it was not detected after testing (vector).}
\item{ProbThisLot}{Individual probability that this lot is still in. It is the complement of the probability
that this specific lot was detected (vector).}
\item{D}{Mean probability of detecting contaminated lots (scalar).}
}
}
\description{
\code{\link[=fvTesting]{fvTesting()}} is a function that simulates the microbiological testing of food unit samples taken from a lot.
Lots are subjected, at a given probability (\code{pLotTested}), to sampling and testing, according to a sampling plan defined by the user.
The algorithm allows for two-class or three-class sampling plans, although the most commonly used for \emph{L. monocytogenes} is the former one:
\code{n} (sample size), \code{c} (maximum level accepted), \code{w} (weight used in the enrichment essay). By performing a bootstrapping on the sampling and testing scheme,
the algorithm estimates that probability of detecting contaminated lots, and the probability that individual lots are contaminated given that they were not detected
as positive during testing. Contaminated lots detected after testing are not removed from the matrix, and therefore the input matrix \code{N} is returned unchanged.
}
\section{Functions}{
\itemize{
\item \code{sfTesting()}: Testing Seafood

}}
\note{
Although microbiological criteria for \emph{L. monocytogenes} are not intended for non-ready-to-eat foods, this function was
conceived to allow the determination of the performance of sampling plans in frozen blanched vegetables as a
potential risk management strategy. \insertCite{PROFEL2020;textual}{qraLm} explains that quick-frozen vegetables should be analysed
for detection of\emph{L. monocytogenes} within the frame of the verification of the food safety management system.
If the pathogen is detected in 25 g, further enumeration needs to be conducted,
on the same samples, to verify if intermediate set limits are reached (\eqn{< 10 CFU/g}) or not.
}
\examples{
dat <- Lot2LotGen(
  nLots = 50,
  sizeLot = 100,
  unitSize = 500,
  betaAlpha = 0.5112,
  betaBeta = 9.959,
  C0MeanLog = 1.023,
  C0SdLog = 0.3267,
  propVarInter = 0.7
)
# Testing the microbiological criterion for RTE foods of absence of \emph{L. monocytogenes}
# in five samples of 25 g each (two-class sampling plan)
Nf1 <- fvTesting(dat,
  nTested = 5, gTested = 25, MTested = 0, cTested = 0,
  pLotTested = 0.5, unitSize = 500, Se = 0.9, iterSub = NULL
)
# Applying an in-house sampling plan of a maximum of one sample (c=1)
# below the M limit 10 CFU/g from a total of five samples taken from a lot
Nf2 <- fvTesting(dat,
  nTested = 5, gTested = 25, MTested = 10, cTested = 1,
  pLotTested = 0.5, unitSize = 500, Se = 0.9, gTestedEnum = 10, iterSub = 800
)

}
\references{
\insertRef{PROFEL2020}{qraLm}
}
\author{
Regis Pouillot \email{rpouillot.work@gmail.com}
}
\keyword{plan}
\keyword{sampling}
\keyword{testing}
\keyword{within-lot}
