% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lot2LotGen.R
\name{Lot2LotGen}
\alias{Lot2LotGen}
\title{Generation of contaminated lots with variability in prevalence and concentration}
\usage{
Lot2LotGen(
  nLots,
  sizeLot,
  unitSize,
  betaAlpha,
  betaBeta,
  C0MeanLog,
  C0SdLog,
  propVarInter,
  Poisson = FALSE,
  ...
)
}
\arguments{
\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{unitSize}{Weight (\code{g}) of single unit or portion from a lot (scalar).}

\item{betaAlpha}{Parameter \code{alpha} of the beta distribution representing the
between-batch variability in prevalence (scalar).}

\item{betaBeta}{Parameter \code{beta} of the beta distribution representing the
between-batch variability in prevalence (scalar).}

\item{C0MeanLog}{Mean parameter (\verb{log10 CFU/g}) of the normal distribution representing
the total variability in the microbial concentration of contaminated units (scalar or vector).}

\item{C0SdLog}{Standard deviation parameter (\verb{log10 CFU/g}) of the normal distribution
representing the total variability in the microbial concentration of contaminated units (scalar or vector).}

\item{propVarInter}{proportion of the total variance (\eqn{C0\_sdLog10^2}) attributed
to between-lot variance. The remaining (\code{1-propVarInter}) is the proportion of within-lot variance.}

\item{...}{Other options used to control \code{\link[=LotGen]{LotGen()}}}

\item{if}{Poisson is \code{TRUE}, a Poisson distribution is used with parameter lambda = -log(1 - prob) / 25, with prob ~ Beta(alpha, beta).
Note that, in this case \code{C0MeanLog} and \code{C0SdLog} are not used.}
}
\value{
A list of four elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units representing the microbial numbers
in the units or portions from contaminated lots;}
\item{\code{ProbUnitPos}}{Probability that the lot is contaminated (a lot is considered contaminated if at
least one unit or portion is contaminated) (vector);}
\item{\code{P}}{Mean prevalence of contaminated lots (scalar);}
\item{\code{betaGen}}{True prevalence of contaminated units in the lots (sampled from the beta distribution) (vector).}
}
}
\description{
The function \code{\link[=Lot2LotGen]{Lot2LotGen()}} generates a matrix of contaminated lots from information on
the parameters of a beta distribution (\code{alpha}, \code{beta}) representing the between-lot variability in prevalence;
the parameters of a normal distribution (\code{mean}, \code{sd}) representing the total variability in the microbial concentration
in the contaminated units (\verb{log10 CFU/g}); and the proportion of such total variance assigned to between-lot variance.
Every row of the matrix corresponds to a production lot, broken down in units or portions, which are the columns of the matrix.
}
\note{
This function requires that the variability in the prevalence of \emph{L. monocytogenes} between lots is
represented by a beta distribution.
To determine the parameters of such a beta distribution, a beta-binomial distribution was fitted to survey data
(compiled in Table 3.21 of \insertCite{Background2022;textual}{qraLm}, consisting of positive samples and total sample size.
The beta-binomial was fitted through a Markov-Chain Monte Carlo approach, with \code{gamma(0.01, 0.01)} as priors for both parameters,
and assuming that every data source (a published paper) represents a different lot. The parameters of the beta distribution
were determined at: \code{betaAlpha=0.5112} and \code{betaBeta=9.959}.
The mean (\code{C0MeanLog=1.023} log10 CFU/g) and standard deviation (\code{C0SdLog=0.3267} log10\ CFU/g) of the normal distribution
representing the overall variability in \emph{L. monocytogenes} were obtained using data from \insertCite{Jeya2011;textual}{qraLm}
(3.0 MPN/g, 3.6 MPN/g in freshly harvested vegetables), \insertCite{Magdovitz2021;textual}{qraLm} (21, 36 and 75 MPN/g in corn and peas
arriving at the frozen food facility), and \insertCite{Kuan2017;textual}{qraLm} (3.0 MPN/g in bulk carrots at retail).
The algorithm decomposes at the log10 level the total variance in concentration into between-lot variance and within-lot
variance according to the proportion \code{propVarInter}.
}
\examples{

dat <- Lot2LotGen(
  nLots = 50,
  sizeLot = 100,
  unitSize = 500,
  betaAlpha = 0.5112,
  betaBeta = 9.959,
  C0MeanLog = 1.023,
  C0SdLog = 0.3267,
  propVarInter = 0.7
)

print(dat$P)
summary(c(dat$N))

}
\references{
\insertRef{stats}{qraLm}
\insertRef{Background2022}{qraLm}
\insertRef{Jeya2011}{qraLm}
\insertRef{Magdovitz2021}{qraLm}
\insertRef{Kuan2017}{qraLm}
}
\seealso{
Other data generation: 
\code{\link{LotGen}()},
\code{\link{caPrimaryProduction}()}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt} and Regis Pouillot \email{rpouillot.work@gmail.com}
}
\concept{data generation}
\keyword{batches}
\keyword{beta-binomial}
\keyword{between-batch}
\keyword{contamination}
\keyword{numbers}
