% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRiskLot.qraLm.R
\name{plotLotRisk.qraLm}
\alias{plotLotRisk.qraLm}
\title{Generic plot function for risk}
\usage{
plotLotRisk.qraLm(x, ...)
}
\arguments{
\item{x}{qraLm object see \code{\link[=Lot2LotGen]{Lot2LotGen()}}}

\item{...}{optional plot parameters passed to the plot function}
}
\description{
Plot risk distribution per contaminated lot
}
\examples{

prod <- Lot2LotGen(
                   nLots = 1000,
                   sizeLot = 1000,
                   unitSize = 500,
                   betaAlpha = 0.5112,
                   betaBeta = 9.959,
                   C0MeanLog = 1.023,
                   C0SdLog = 0.3267,
                   propVarInter = 0.7
                   )
DRmodel <- "JEMRA"
population <- 2
risk <- DRForModel(prod,
                   model = DRmodel,
                   population = population)
str(risk)
plotLotRisk.qraLm(risk)

}
\author{
Vasco Cadavez
}
