% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLotsECDF.qraLm.R
\name{plotLotsECDF.qraLm}
\alias{plotLotsECDF.qraLm}
\title{Plot the Empirical Cumulative Distribution Function (ECDF) of CFU/g in a contaminated lot}
\usage{
plotLotsECDF.qraLm(x, ...)
}
\arguments{
\item{x}{A \code{qraLm} object containing lotMeans.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{plotly} interactive ECDF plot.
}
\description{
This function plots the ECDF of CFU/g in a contaminated lot using ggplot2 and plotly.
}
\examples{
# Example usage (assuming x is a qraLm object)
# plotLotsECDF.qraLm(x)
}
