% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUnitsECDF.qraLm.R
\name{plotUnitsECDF.qraLm}
\alias{plotUnitsECDF.qraLm}
\title{plotUnitsECDF Generic plot function to plot the ECDF}
\usage{
plotUnitsECDF.qraLm(x, ...)
}
\arguments{
\item{x}{qraLm object see \code{\link[=Lot2LotGen]{Lot2LotGen()}}}

\item{...}{Optional plot parameters passed to the function}
}
\description{
Plot ECDF curve
}
\examples{

prod <- Lot2LotGen(
                   nLots = 1000,
                   sizeLot = 1000,
                   unitSize = 500,
                   betaAlpha = 0.5112,
                   betaBeta = 9.959,
                   C0MeanLog = 1.023,
                   C0SdLog = 0.3267,
                   propVarInter = 0.7
                   )
prod1 <- Lot2LotGen(
                    nLots = 1000,
                    sizeLot = 1000,
                    unitSize = 500,
                    betaAlpha = 0.9112,
                    betaBeta = 2.959,
                    C0MeanLog = 2.023,
                    C0SdLog = 0.5267,
                    propVarInter = 0.7
                    )
str(prod1)
plotUnitsECDF.qraLm(prod1)

p1 <- plotUnitsECDF.qraLm(prod)
p2 <- plotUnitsECDF.qraLm(prod1)
par(mfrow = c(1, 2))
p1
p2
par(mfrow = c(1, 1)) # reset this parameter

}
\author{
Vasco Cadavez
}
