% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qraLm.R
\docType{package}
\name{qraLm}
\alias{qraLm}
\alias{qraLm-package}
\title{qraLm: An R package for the quantitative risk assessment of Listeria monocytogenes in foods}
\description{
Functions for the quantitative risk assessment of Listeria monocytogenes in foods
}
\section{qraLm functions}{

The qraLm functions ...
}

