% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskEurope.R
\name{RiskEurope}
\alias{RiskEurope}
\title{Risk Europe}
\usage{
RiskEurope(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  ProbUnitPos,
  ProbThisLot,
  Virulence = ""
)
}
\arguments{
\item{data}{A list of four elements see \code{\link[=Lot2LotGen]{Lot2LotGen()}} function.}

\item{nLots}{see \code{\link[=Lot2LotGen]{Lot2LotGen()}} function.}

\item{sizeLot}{see \code{\link[=Lot2LotGen]{Lot2LotGen()}} function.}

\item{ProbUnitPos}{is}

\item{ProbThisLot}{is}

\item{Virulence}{is}
}
\description{
The function \code{\link[=RiskEurope]{RiskEurope()}}  ..
}
\note{
notas
}
\examples{
library(qraLm)
sizeLot <- 5
nLots <- 3
N <- matrix(10^rnorm(sizeLot * nLots, 3, 2), ncol = sizeLot, nrow = nLots)
N[1, ] <- 0
ProbUnitPos <- rep(0.5, nLots)
ProbThisLot <- c(.1, .2, .3)
Virulence <- ""
# error Error in dimnames(x) <- dn
# :length of 'dimnames' [2] not equal to array extent
# RiskEurope(N, ProbUnitPos, ProbThisLot, "")
}
