% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Shiny application for \code{qraLm} package}
\usage{
runExample()
}
\description{
Shiny application to simulate quantitative risk assessment for \emph{L. monocytogenes}
in Frozen Vegetables, Diced RTE Cantaloupe and Cold-smoked RTE Fish.
}
\references{
\insertRef{stats}{qraLm}
}
\author{
Vasco Cadavez \email{vcadavez@ipb.pt}
}
