% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfBriningCC.R
\name{sfBriningCC}
\alias{sfBriningCC}
\title{Internal contamination of fish fillets during brine injection}
\usage{
sfBriningCC(
  data = list(),
  pccBrine,
  volInjMin,
  volInjMode,
  volInjMax,
  concBrineMin,
  concBrineMode,
  concBrineMax
)
}
\arguments{
\item{data}{a list of
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of \verb{size nLots} lots by \code{sizeLot} units containing the numbers of \emph{L. monocytogenes} in/on fish fillets;}
\item{\code{P}}{Mean prevalence of contaminated lots (scalar);}
\item{\code{ProbUnitPos}}{Probability of individual lots being contaminated (vector).}
}}

\item{pccBrine}{Probability that the brine solution is contaminated with \emph{L. monocytogenes}  (scalar).}

\item{volInjMin}{(\code{ml}) minimum volume of brine solution injected in a fish fillet (scalar).}

\item{volInjMode}{(\code{ml}) most likely volume of brine solution injected in a fish fillet (scalar).}

\item{volInjMax}{(\code{ml}) maximum volume of brine solution injected in a fish fillet (scalar).}

\item{concBrineMin}{(\code{CFU/ml}) minimum concentration of \emph{L. monocytogenes} in contaminated brine solution (scalar).}

\item{concBrineMode}{(\code{CFU/ml}) most likely concentration of \emph{L. monocytogenes} in contaminated brine solution (scalar).}

\item{concBrineMax}{(\code{CFU/ml}) maximum concentration of \emph{L. monocytogenes} in contaminated brine solution (scalar).}
}
\value{
A list of three elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units containing the numbers of \emph{L. monocytogenes} in brined fish;}
\item{\code{ProbUnitPos}}{Probability of individual lots being contaminated after brining (vector);}
\item{\code{P}}{Mean prevalence of contaminated lots after brining (scalar).}
}
}
\description{
The function \code{\link[=sfBriningCC]{sfBriningCC()}} simulates the potential internal contamination of fish fillets during brining by injection
of salt solution. The cross-contamination algorithm accounts for four possible scenarios:
\enumerate{
\item cross-contamination occurring in lots already contaminated;
\item cross-contamination occurring in lots that were not contaminated;
\item no additional contamination occurring in lots already contaminated; and
\item no cross-contamination occurring in lots that were not contaminated. Probabilities of occurrence of the four events are computed.
}
The mean volume of brine injected per fillet (\code{V_inj}) and the mean concentration of \emph{L. monocytogenes} in brine (\code{N_brine}) are
allowed to vary from lot to lot; and, to this effect, their values are sampled from Pert distributions with parameters \code{volInjMin}, \code{volInjMode},
\code{volInjMax}; and \code{concBrineMin}, \code{concBrineMode}, \code{concBrineMax}, respectively. To estimate the number of cells transferred to the
fish fillet, it is assumed that the cells follow a Poisson distribution in the volume of brine.
}
\note{
The suggested value of \eqn{probCCDice\_brine=0.135} is taken from \insertCite{Gudbjornsdottir2005;textual}{qraLm} and
\insertCite{Gudbjornsdottir2004;textual}{qraLm}, who analysed the presence of \emph{L. monocytogenes} in brining solution, detecting
3 positive samples out of 14, and 2 positive samples out of 23, respectively.
The parameters of the distributions about mean volume of brine solution injected in a fish fillet and mean concentration of \emph{L. monocytogenes} in brine
must be defined by the user and/or tested in scenarios.
}
\examples{
nLots <- 100
sizeLot <- 50
pccBrine <- 5 / 37
volInjMin <- 0.5
volInjMode <- 1.2
volInjMax <- 4.5
concBrineMin <- 0
concBrineMode <- 3
concBrineMax <- 40
ProbUnitPos <- rep(0.105, nLots)

dat <- list(
  N = matrix(rpois(nLots * sizeLot, 16),
    nrow = nLots,
    ncol = sizeLot
  ),
  P = 0.22,
  ProbUnitPos = ProbUnitPos,
  nLots = 100,
  sizeLot = 50
)
Nf <- sfBriningCC(
  dat,
  pccBrine,
  volInjMin, volInjMode, volInjMax,
  concBrineMin, concBrineMode, concBrineMax
)
hist(Nf$N)
}
\references{
\insertRef{Gudbjornsdottir2004}{qraLm}
\insertRef{Gudbjornsdottir2005}{qraLm}
\insertRef{mc2d}{qraLm}
\insertRef{stats}{qraLm}
\insertRef{iRisk2021}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt} and Regis Pouillot \email{rpouillot.work@gmail.com}
}
\keyword{brine}
\keyword{contamination}
\keyword{injection}
\keyword{internal}
\keyword{solution}
\keyword{transfer}
