% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfCharacteristics.R
\name{sfCharacteristics}
\alias{sfCharacteristics}
\title{Sampling of the RTE smoked fish characteristics}
\usage{
sfCharacteristics(
  nLots = data$nLots,
  awminSF = NULL,
  awmodeSF = NULL,
  awmaxSF = NULL,
  NaClminSF = 1.5,
  NaClmodeSF = 3.4,
  NaClmaxSF = 5.3,
  PminSF = 5,
  PmodeSF = 10,
  PmaxSF = 22,
  pHminSF = 5.8,
  pHmodeSF = 6.1,
  pHmaxSF = 6.5,
  CO2equilibriumminSF = 0.25,
  CO2equilibriummodeSF = 0.25,
  CO2equilibriummaxSF = 0.3,
  NITminSF = 0,
  NITmodeSF = 0,
  NITmaxSF = 0,
  aaWphminSF = 0,
  aaWphmodeSF = 0,
  aaWphmaxSF = 0,
  baWphminSF = 0,
  baWphmodeSF = 0,
  baWphmaxSF = 0,
  caWphminSF = 0,
  caWphmodeSF = 0,
  caWphmaxSF = 0,
  daWphminSF = 500,
  daWphmodeSF = 1500,
  daWphmaxSF = 1900,
  laWphminSF = 6000,
  laWphmodeSF = 12000,
  laWphmaxSF = 28000,
  saWphminSF = 0,
  saWphmodeSF = 0,
  saWphmaxSF = 0
)
}
\arguments{
\item{nLots}{Number of lots or size of the Monte Carlo simulation}

\item{awminSF}{Minimum water activity of RTE. If null,it will be calculated from NaCl concentration.}

\item{awmodeSF}{Mode of water activity of RTE.}

\item{awmaxSF}{Maximum water activity of RTE.}

\item{NaClminSF}{(\%) Minimum NaCl of RTE}

\item{NaClmodeSF}{(\%) Mode of NaCl of RTE}

\item{NaClmaxSF}{(\%) Maximum NaCl of RTE}

\item{PminSF}{(ppm) Minimum phenol concentration in RTE}

\item{PmodeSF}{(ppm) Mode of phenol concentration in RTE}

\item{PmaxSF}{(ppm) Maximum phenol concentration in RTE}

\item{pHminSF}{Minimum pH of RTE}

\item{pHmodeSF}{Mode of pH of RTE}

\item{pHmaxSF}{Maximum pH of RTE}

\item{CO2equilibriumminSF}{(proportion) Minimum \code{CO_2} concentration in atmosphere in RTE package (e.g. .0.25)}

\item{CO2equilibriummodeSF}{(proportion) Mode of \code{CO_2} concentration in atmosphere in RTE package}

\item{CO2equilibriummaxSF}{(proportion) Maximum \code{CO_2} concentration in atmosphere in RTE package}

\item{NITminSF}{(ppm) Minimum nitrites concentration in RTE}

\item{NITmodeSF}{(ppm) Mode of nitrites concentration in RTE}

\item{NITmaxSF}{(ppm) Maximum nitrites concentration in RTE}

\item{aaWphminSF}{(ppm) Minimum acetic acid concentration in RTE}

\item{aaWphmodeSF}{(ppm) Mode of acetic acid concentration in RTE}

\item{aaWphmaxSF}{(ppm) Maximum acetic acid concentration in RTE}

\item{baWphminSF}{(ppm) Minimum benzoic acid concentration in RTE}

\item{baWphmodeSF}{(ppm) Mode benzoic acid concentration in RTE}

\item{baWphmaxSF}{(ppm) Maximum benzoic acid concentration in RTE}

\item{caWphminSF}{(ppm) Minimum citric acid concentration in RTE}

\item{caWphmodeSF}{(ppm) Mode of citric acid concentration in RTE}

\item{caWphmaxSF}{(ppm) Maximum citric acid concentration in RTE}

\item{daWphminSF}{(ppm) Minimum diacetate concentration in RTE}

\item{daWphmodeSF}{(ppm) Mode of diacetate concentration in RTE}

\item{daWphmaxSF}{(ppm) Maximum diacetate concentration in RTE}

\item{laWphminSF}{(ppm) Minimum lactic acid concentration in RTE}

\item{laWphmodeSF}{(ppm) Mode of lactic acid concentration in RTE}

\item{laWphmaxSF}{(ppm) Maximum lactic acid concentration in RTE}

\item{saWphminSF}{(ppm) Minimum sorbic acid concentration in RTE}

\item{saWphmodeSF}{(ppm) Mode of sorbic acid concentration in RTE}

\item{saWphmaxSF}{(ppm) Maximum sorbic acid concentration in RTE}
}
\value{
A list containing vectors \code{aw}, \code{NaCl}, \code{pH}, P, \code{CO2equi}, \code{NIT}, \code{aaWph}, \code{baWph},
\code{caWph}, \code{daWph}, \code{laWph}, \code{saWph} to be used in the growth functions \code{\link[=sfMejlholmDalgaard]{sfMejlholmDalgaard()}} and
\code{\link[=sfMejlholmDalgaardLAB]{sfMejlholmDalgaardLAB()}} for estimating specific growth rates of \emph{L. monocytogenes} and lactic
acid bacteria, respectively, in RTE seafood.
}
\description{
The function \code{\link[=sfCharacteristics]{sfCharacteristics()}} samples from multiple Pert distributions the
intrinsic and extrinsic characteristics of RTE smoked seafood, which affects microbial growth; namely,
water activity, pH, contents of salt, phenol, nitrites, acetic acid, benzoic acid, citric acid, diacetate,
lactic acid and sorbic acid, and concentration of \code{CO_2} in equilibrium in the package.
Since sampling is carried out at lot level; all the RTE seafood units of a lot have the same characteristics.
}
\note{
All parameters end with \code{SF} not to be confounded with growth parameters. If parameters for water activity \code{aw} are \code{NULL},
\code{aw} will be evaluated from NaCl concentration in the growth functions. The default parameters of the Pert distributions describing
variability in \code{pH}, contents of \code{salt}, \code{phenol}, \code{diacetate}, \verb{lactic acid}, and concentration of \code{CO_2} in equilibrium for smoked seafood,
were taken from \insertCite{Background2022;textual}{qraLm} (Table 3.25) and \insertCite{EFSA2017;textual}{qraLm}.
}
\examples{
N <- matrix(round(10^rnorm(100 * 100, 0, 3)),
  ncol = 100, nrow = 100
)

dat <- list(N = N, nLots = 1000)
dat <- sfCharacteristics(dat)
str(dat)

}
\references{
\insertRef{Background2022}{qraLm}
\insertRef{EFSA2017}{qraLm}
\insertRef{mc2d}{qraLm}
}
\author{
Regis Pouillot \email{rpouillot.work@gmail.com}
}
\keyword{seafood}
