% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfColdChain.R
\name{sfColdChain}
\alias{sfColdChain}
\title{Cold chain logistics for ready-to-eat seafood}
\usage{
sfColdChain(
  data = list(),
  RTE,
  nLots = NULL,
  sizeLot = NULL,
  unitSize = NULL,
  tempMin = 0.288,
  tempMode = 4.6,
  tempMax = 8.912,
  timeMin = 12,
  timeMode = 144,
  timeMax = 360,
  variability = "lot",
  corTimeTemp = 0,
  N0LABmin = 1.5,
  N0LABmode = 2.78,
  N0LABmax = 4.1,
  intralotSdN0LAB = 0,
  lnQ0LABmin = -12,
  lnQ0LABmode = -2.73,
  lnQ0LABmax = 1.26,
  MPDLABmin = 8,
  MPDLABmode = 8.5,
  MPDLABmax = 9,
  MPDLmmin = 6.6,
  MPDLmmode = 7.36,
  MPDLmmax = 8.2,
  mumaxrefLm = 0.419,
  TminLm = -2.83,
  TrefLm = 25,
  awminLm = 0.923,
  pHminLm = 4.97,
  pheMaxLm = 32,
  NITmaxLm = 350,
  CO2maxLm = 3140,
  micLACuLm = 3.79,
  micDACuLm = 4.8,
  micAACuLm = 10.3,
  micBACuLm = 0.349,
  micCACuLm = 2.119,
  micSACuLm = 1.896,
  mumaxrefLAB = 0.583,
  TminLAB = -5.25,
  TrefLAB = 25,
  awminLAB = 0.928,
  pHminLAB = 4.24,
  pheMaxLAB = 40.3,
  NITmaxLAB = 2780,
  CO2maxLAB = 6691,
  micLACuLAB = 12,
  micDACuLAB = 33.3,
  micAACuLAB = 10.3,
  micBACuLAB = 1.51,
  micCACuLAB = 10.3,
  micSACuLAB = 12.6,
  gamma = 1,
  lim = 1,
  step = 1
)
}
\arguments{
\item{data}{a list of:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix containing the numbers of \emph{L. monocytogenes} in packs of RTE seafood at the end of transport to retail, from contaminated lots.}
\item{\code{P}}{Prevalence of contaminated lots (scalar).}
\item{\code{N0LAB}}{(\code{CFU}) Optional: A matrix containing the numbers of LAB in packs of RTE seafood at the end of transport to retail, from contaminated lots. If missing, it will be generated from default parameters.}
\item{\code{lnQ0Lm}}{Optional: A matrix containing the natural logarithm of q0 for \emph{L. monocytogenes} in packs of RTE seafood. If missing, it will be generated from default parameters.}
\item{\code{lnQ0LAB}}{Optional: A matrix containing the natural logarithm of q0 for LAB in packs of RTE seafood. If missing, it will be generated from default parameters.}
\item{\code{MPDLm}}{(\verb{log10 CFU/g}) Optional: A matrix containing the maximum population density of \emph{L. monocytogenes} in packs of RTE seafood. If missing, it will be generated from default parameters.}
\item{\code{MPDLAB}}{(\verb{log10 CFU/g}) Optional: A matrix containing the log10 of MPD for LAB in packs of RTE seafood. If missing, it will be generated from default parameters.}
}}

\item{RTE}{Intrinsic and extrinsic characteristics of the RTE seafood product, provided using the \code{\link{sfCharacteristics}} function}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{unitSize}{(\code{g}) Weight of the contents of the RTE seafood product}

\item{tempMin}{(\eqn{^\circ C}) Minimum storage temperature of the RTE seafood product (scalar). Will be used in a Pert distribution.}

\item{tempMode}{(\eqn{^\circ C}) Mode of storage temperature of the RTE seafood product (scalar).}

\item{tempMax}{(\eqn{^\circ C}) Maximum storage temperature of the RTE seafood product (scalar).}

\item{timeMin}{(\code{h}) Minimum storage time of the RTE seafood product (scalar). Will be used in a Pert distribution.}

\item{timeMode}{(\code{h}) Mode of storage time of the RTE seafood product (scalar).}

\item{timeMax}{(\code{h}) Maximum storage time of the RTE seafood product (scalar).}

\item{variability}{level of variability for time and temperature: "lot", "column" or "portion".}

\item{corTimeTemp}{define}

\item{N0LABmin}{(\verb{log10 CFU}) Minimum value of the Pert distribution representing variability in the numbers of LAB at \code{time=0} (scalar). Will be then used as a power of 10.}

\item{N0LABmode}{(\verb{log10 CFU}) Mode value of the Pert distribution representing variability in the numbers of LAB at \code{time=0} (scalar).}

\item{N0LABmax}{(\verb{log10 CFU}) Maximum value of the Pert distribution representing variability in the numbers of LAB at \code{time=0} (scalar).}

\item{intralotSdN0LAB}{(\verb{log10 CFU}) Intra-lot standard deviation to produce variability in the numbers of LAB (scalar).}

\item{lnQ0LABmin}{Minimum value of the Pert distribution representing variability in the natural logarithm of the parameter \code{q0} for LAB (scalar).}

\item{lnQ0LABmode}{Mode value of the Pert distribution representing variability in the natural logarithm of the parameter \code{q0} for LAB (scalar).}

\item{lnQ0LABmax}{Maximum value of the Pert distribution representing variability in the natural logarithm of the parameter \code{q0} for LAB (scalar).}

\item{MPDLABmin}{(\verb{log10 CFU/g}) Minimum value of the Pert distribution representing the variability in the maximum population density of LAB (scalar).}

\item{MPDLABmode}{(\verb{log10 CFU/g}) Mode value of the Pert distribution representing the variability in the maximum population density of LAB (scalar).}

\item{MPDLABmax}{(\verb{log10 CFU/g}) Maximum value of the Pert distribution representing the variability in the maximum population density of LAB (scalar).}

\item{MPDLmmin}{(\verb{log10 CFU/g}) Minimum value of the Pert distribution representing the variability in the maximum population density of \emph{L. monocytogenes} (scalar).}

\item{MPDLmmode}{(\verb{log10 CFU/g}) Mode value of the Pert distribution representing the variability in the maximum population density of \emph{L. monocytogenes} (scalar).}

\item{MPDLmmax}{(\verb{log10 CFU/g}) Maximum value of the Pert distribution representing the variability in the maximum population density of \emph{L. monocytogenes} (scalar).}

\item{mumaxrefLm}{(\code{1/h}) Maximum growth rate of \emph{L. monocytogenes} at the reference temperature \code{TrefLm} (scalar).}

\item{TminLm}{(\eqn{^\circ C}) Minimum temperature for growth of \emph{L. monocytogenes} (scalar).}

\item{TrefLm}{(\eqn{^\circ C}) Reference temperature for \emph{L. monocytogenes} (scalar).}

\item{awminLm}{Minimum water activity for growth of \emph{L. monocytogenes} (scalar).}

\item{pHminLm}{Minimum pH for growth of \emph{L. monocytogenes} (scalar).}

\item{pheMaxLm}{(\code{ppm}) Minimum inhibitory concentration of phenol for \emph{L. monocytogenes} (scalar).}

\item{NITmaxLm}{(\code{ppm}) Minimum inhibitory concentration of nitrites for \emph{L. monocytogenes} (scalar).}

\item{CO2maxLm}{(\code{ppm}) Minimum inhibitory concentration of CO2 for \emph{L. monocytogenes} (scalar).}

\item{micLACuLm}{(\code{mM}) Minimum inhibitory concentration of undissociated lactic acid for \emph{L. monocytogenes} (scalar).}

\item{micDACuLm}{(\code{mM}) Minimum inhibitory concentration of undissociated diacetate for \emph{L. monocytogenes} (scalar).}

\item{micAACuLm}{(\code{mM}) Minimum inhibitory concentration of undissociated acetic acid for \emph{L. monocytogenes} (scalar).}

\item{micBACuLm}{(\code{mM}) Minimum inhibitory concentration of undissociated benzoic acid for \emph{L. monocytogenes} (scalar).}

\item{micCACuLm}{(\code{mM}) Minimum inhibitory concentration of undissociated citric acid for \emph{L. monocytogenes} (scalar).}

\item{micSACuLm}{(\code{mM}) Minimum inhibitory concentration of undissociated sorbic acid for \emph{L. monocytogenes} (scalar).}

\item{mumaxrefLAB}{(\code{1/h}) Maximum growth rate of LAB at the reference temperature \code{TrefLAB} (scalar).}

\item{TminLAB}{(\eqn{^\circ C}) Minimum temperature for growth of LAB (scalar).}

\item{TrefLAB}{(\eqn{^\circ C}) Reference temperature for LAB (scalar).}

\item{awminLAB}{Minimum water activity for growth of LAB (scalar).}

\item{pHminLAB}{Minimum pH for growth of LAB (scalar).}

\item{pheMaxLAB}{(\code{ppm}) Minimum inhibitory concentration of phenol for LAB (scalar).}

\item{NITmaxLAB}{(\code{ppm}) Minimum inhibitory concentration of nitrites for LAB (scalar).}

\item{CO2maxLAB}{(\code{ppm}) Minimum inhibitory concentration of CO2 for LAB (scalar).}

\item{micLACuLAB}{(\code{Mm}) Minimum inhibitory concentration of undissociated lactic acid for LAB (scalar).}

\item{micDACuLAB}{(\code{Mm}) Minimum inhibitory concentration of undissociated diacetate for LAB (scalar).}

\item{micAACuLAB}{(\code{Mm}) Minimum inhibitory concentration of undissociated acetic acid for LAB (scalar).}

\item{micBACuLAB}{(\code{Mm}) Minimum inhibitory concentration of undissociated benzoic acid for LAB (scalar).}

\item{micCACuLAB}{(\code{Mm}) Minimum inhibitory concentration of undissociated citric acid for LAB (scalar).}

\item{micSACuLAB}{(\code{Mm}) Minimum inhibitory concentration of undissociated sorbic acid for LAB (scalar).}

\item{gamma}{Gamma parameter for the Jameson effect function. If 1, converges to classical Jameson effect (scalar).}

\item{lim}{(\verb{log10 CFU/g}) Limit to consider interactions in the Jameson effect (scalar).}

\item{step}{(\code{h}) Integration step (scalar).}
}
\value{
A list of six elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix containing the numbers of \emph{L. monocytogenes} in packs of RTE seafood after storage,
from contaminated lots;}
\item{\code{P}}{Prevalence of RTE seafood lots contaminated with \emph{L. monocytogenes} (scalar);}
\item{\code{N0LAB}}{(\code{CFU}) A matrix containing the numbers of LAB in packs of RTE seafood;}
\item{\code{lnQ0Lm}}{A matrix containing the natural logarithm of \code{q0} of \emph{L. monocytogenes} in packs of RTE seafood;}
\item{\code{lnQ0LAB}}{A matrix containing the natural logarithm of \code{q0} of LAB in packs of RTE seafood;}
\item{\code{SF}}{An object containing the intrinsic and extrinsic characteristics of the RTE seafood product.}
}
}
\description{
The function \code{\link[=sfColdChain]{sfColdChain()}} simulates the growth of \emph{L. monocytogenes}, as affected by lactic acid bacteria (LAB) populations,
in RTE seafood during the various stages of the cold chain logistics, including transportation to retail, display at retail and transportation
to home; and relies on the growth kinetic functions sfMejlholmDalgaard, \code{\link[=sfMejlholmDalgaardLAB]{sfMejlholmDalgaardLAB()}} and \code{\link[=sfGrowthJameson]{sfGrowthJameson()}}.
The function \code{\link[=sfColdChain]{sfColdChain()}} can be used for home storage as well, by changing only the parameters defining the time and
temperature Pert distributions. For every lot, the algorithm samples time and temperature from Pert distributions defined by the
parameters \code{timeMin}, \code{timeMode}, \code{timeMax}, and \code{tempMin}, \code{tempMode}, \code{tempMax}, respectively, provided by the user.
The specific growth rates for \emph{L. monocytogenes} and LAB are determined evaluating the intrinsic and extrinsic characteristics
of the seafood product (sampled from the function \code{\link[=sfCharacteristics]{sfCharacteristics()}}) in the Mejholm and Daalgard's existing secondary
models for both bacterial groups (\insertCite{Mejlholm2009}{qraLm}, \insertCite{Mejlholm2010}{qraLm}, \insertCite{Mejlholm2013}{qraLm}
and \insertCite{Mejlholm2015}{qraLm}). The parameters of the two growth rate models related to reference growth rate, temperature,
water activity, pH and concentrations of phenol, nitrites, CO2 at equilibrium, undissociated lactic acid, undissociated diacetate,
undissociated acetic acid, undissociated benzoic acid, undissociated citric and undissociated sorbic acid, for \emph{L. monocytogenes}
and lactic acid bacteria, are set by default to those fitted and validated by \insertCite{Mejlholm2010;textual}{qraLm} and
\insertCite{Mejlholm2013;textual}{qraLm}, respectively.
However, any of the parameters of the growth rate models can be altered or updated. The numbers of \emph{L. monocytogenes} in the
RTE seafood product, as affected by LAB, are then estimated applying the extended Jameson-effect competition model published
in \insertCite{Gimenez2004;textual}{qraLm} using the interaction \code{gamma} parameter proposed by \insertCite{Moller2013;textual}{qraLm}.
}
\note{
This function can be used multiple times. The microbial kinetic parameters \code{MPD} and \code{ln_q0} for \emph{L. monocytogenes} and
LAB are sampled only once. Information for building the variability distributions about retail time and temperature was obtained from
\insertCite{Pouillot2007;textual}{qraLm} and \insertCite{FDAFSIS2003;textual}{qraLm}.
For Home storage, the Pert distribution for temperature can be defined using the following parameters: \code{tempMin = qnorm(0.025, 7, 3)},
\code{tempMode = 7}, and \code{tempMax = qnorm(0.975, 7, 3)}, which are reformulated from the \code{normal(7.0, 3.0)} used in
\insertCite{Pouillot2007;textual}{qraLm}.
The Pert distribution for home storage time can be defined using the following parameters:
\code{timeMin = qweibull(0.025,shape = 1.14, scale = 18.39) * 24}
\code{timeMode = 18.39*((1.14-1)/1.14)^(1/1.14) * 24}
\code{timeMax = qweibull(0.975,shape = 1.14, scale = 18.39) * 24}
which are reformulated from the \code{Weibull(1.14, 18.39)} days used in \insertCite{Endrikat2010;textual}{qraLm}.
}
\examples{

sizeLot <- 100
size_mc <- 100
N <- matrix(round(10^rnorm(size_mc * sizeLot, 0, 3)),
            ncol = sizeLot,
            nrow = size_mc
            )

dat <- list(N = N, unitSize = 500)
RTE <- sfCharacteristics(size_mc)
str(RTE)
dat1 <- sfColdChain(dat,
                    RTE = RTE
                    )
str(dat1)


}
\references{
\insertRef{Endrikat2010}{qraLm}
\insertRef{extraDistr}{qraLm}
\insertRef{FDAFSIS2003}{qraLm}
\insertRef{Gimenez2004}{qraLm}
\insertRef{mc2d}{qraLm}
\insertRef{Mejlholm2009}{qraLm}
\insertRef{Mejlholm2010}{qraLm}
\insertRef{Mejlholm2013}{qraLm}
\insertRef{Mejlholm2015}{qraLm}
\insertRef{Moller2013}{qraLm}
\insertRef{Pouillot2007}{qraLm}
\insertRef{stats}{qraLm}
}
\author{
Regis Pouillot \email{rpouillot.work@gmail.com}
}
\keyword{Baranyi}
\keyword{competition}
\keyword{growth}
\keyword{lag-phase}
\keyword{logistics}
\keyword{stochastic}
