% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGrowthJameson.R
\name{sfGrowthJameson}
\alias{sfGrowthJameson}
\title{Growth model of \emph{L. monocytogenes} in RTE seafood including background flora competition}
\usage{
sfGrowthJameson(
  time,
  N0Lm,
  N0LAB,
  q0Lm,
  q0LAB,
  muLm,
  muLAB,
  MPDLm,
  MPDLAB,
  unitSize,
  gamma = 1,
  lim = 1,
  step = 1
)
}
\arguments{
\item{time}{(h) Storage time of a pack of RTE seafood}

\item{N0Lm}{(CFU) Numbers of \emph{L. monocytogenes} in RTE seafood at time=0}

\item{N0LAB}{(CFU) Numbers of LAB in RTE seafood at time=0}

\item{q0Lm}{Initial parameter Q of the Baranyi and Roberts' model related to lag phase for \emph{L. monocytogenes} at time=0}

\item{q0LAB}{Initial parameter Q of the Baranyi and Roberts' model related to lag phase for LAB at time=0}

\item{muLm}{(h-1) Growth rate of \emph{L. monocytogenes} in RTE seafood}

\item{muLAB}{(h-1) Growth rate of LAB in RTE seafood}

\item{MPDLm}{(log10 CFU/g) Maximum population density of \emph{L. monocytogenes} in RTE seafood}

\item{MPDLAB}{(log10 CFU/g) Maximum population density of LAB in RTE seafood}

\item{unitSize}{(g) Weight of the contents of an RTE seafood pack (for MPD)}

\item{gamma}{Interaction parameter of the competition model. Note: if gamma=1, population 1 stops when population 2 reaches MPD.
If gamma < 1, population 1 still increases when population 2 reaches MPD. If gamma > 1, population 1 decreases when population 2 reaches MPD.}

\item{lim}{= 1 A parameter indicating when the competition should be considered as meaningful.}

\item{step}{= 1 Integration step.}
}
\value{
a list of four elements:
\describe{
\item{N1Lm}{(CFU) Numbers of \emph{L. monocytogenes} at the end of storage;}
\item{N1LAB}{(CFU) Numbers of LAB at the end of storage;}
\item{lnQtLm}{Natural logarithm of q0 for \emph{L. monocytogenes} at the end of storage;}
\item{lnQtLAB}{Natural logarithm of q0 for LAB at the end of storage.}
}
}
\description{
The \code{\link[=sfGrowthJameson]{sfGrowthJameson()}} function is used to simulate the growth of \emph{L. monocytogenes} in RTE seafood, as affected by lactic acid bacteria (LAB),
according to the Jameson-effect competition model published in \insertCite{Gimenez2004;textual}{qraLm}.
}
\note{
For sake of speed, the function first evaluates a classical Baranyi and Roberts model without interactions.
If any of the two population reaches a value greater than \eqn{10^{(MPD - lim)}} CFU, an expanded version of the Jameson-effect model
by \insertCite{Gimenez2004;textual}{qraLm} is used. This model includes the parameter gamma proposed by \insertCite{Moller2013;textual}{qraLm}.
The function is vectorized for all parameters.
}
\examples{
library(qraLm)
iter <- 1000
set.seed(666)
vtime <- 200
vN0Lm <- 10^runif(iter, 0, 3)
vN0LAB <- 10^runif(iter, 0, 3)
vq0Lm <- exp(runif(iter, 0, 1))
vq0LAB <- exp(runif(iter, 0, 1))
vmuLm <- runif(iter, 0, 0.1)
vmuLAB <- runif(iter, 0, 0.1)
sfGrowthJameson(
  time = vtime, N0Lm = vN0Lm, N0LAB = vN0LAB,
  q0Lm = vq0Lm, q0LAB = vq0LAB,
  muLm = vmuLm, muLAB = vmuLAB,
  MPDLm = 8, MPDLAB = 7,
  unitSize = 100,
  gamma = 0.8
)

# This is not very efficient (201 integrations), but it works.
Res <- sfGrowthJameson(
  time = 0:200, N0Lm = rep(1, 201), N0LAB = rep(100, 201),
  q0Lm = rep(1, 201), q0LAB = rep(1, 201),
  muLm = rep(.1, 201), muLAB = rep(.3, 201),
  MPDLm = rep(8, 201), MPDLAB = rep(7, 201),
  unitSize = rep(100, 201),
  gamma = rep(0.8, 201)
)
plot(0:200, log10(Res$N1Lm), type = "l", ylim = c(0, 9), xlab = "time", ylab = "log10 cfu")
points(0:200, log10(Res$N1LAB), type = "l", col = "blue")
legend(x = 0, y = 8.5, legend = c("Listeria", "Food flora"), col = c("black", "blue"), lty = 1)

}
\references{
\insertRef{Gimenez2004}{qraLm}

\insertRef{Moller2013}{qraLm}
}
\author{
Regis Pouillot \email{rpouillot.work@gmail.com} and Laurent Guillier
}
\keyword{competition}
\keyword{inhibition}
\keyword{interaction}
