% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfGrowthLPD.R
\name{sfGrowthLPD}
\alias{sfGrowthLPD}
\title{Growth of \emph{L. monocytogenes} in raw fish}
\usage{
sfGrowthLPD(
  data = list(),
  unitSize = NULL,
  MPD = NULL,
  Temp,
  time,
  aLM = 0.0581,
  Tmin = 1.3,
  ALM = 0.84,
  mLM = 1.11,
  workDone = 0
)
}
\arguments{
\item{data}{See \code{\link[=Lot2LotGen]{Lot2LotGen()}} function.}

\item{unitSize}{(g) Weight of a raw fish unit (whole or fillet).}

\item{MPD}{(log10 CFU/g) Maximum population density of \emph{L. monocytogenes} in raw fish, taken from \insertCite{Jia2020;textual}{qraLm}.}

\item{Temp}{(\eqn{^\circ C}) Temperature of raw fish fillet during holding time or storage.}

\item{time}{(h) Length of holding time or storage.}

\item{aLM}{Coefficient of the secondary model for growth rate, taken from \insertCite{Jia2020;textual}{qraLm}.}

\item{Tmin}{(\eqn{^\circ C}) Minimum temperature for growth of \emph{L. monocytogenes} in raw fish, taken from \insertCite{Jia2020;textual}{qraLm}.}

\item{ALM}{Coefficient of the model for lag phase duration in hours, taken from \insertCite{Jia2020;textual}{qraLm}.}

\item{mLM}{Exponent of the model for lag phase duration in hours, taken from \insertCite{Jia2020;textual}{qraLm}.}

\item{workDone}{Work done during the stages prior to the current storage stage. Should be set to zero if there was no previous
storage phase accounted for.}
}
\value{
A list of two elements:
\describe{
\item{N}{(CFU) Numbers of \emph{L. monocytogenes} in fish fillet after storage;}
\item{workDone}{Work done during the previous stages including the current storage stage.}
}
}
\description{
The function \code{\link[=sfGrowthLPD]{sfGrowthLPD()}} simulates the growth of \emph{L. monocytogenes} in raw fish (salmon) considering lag phase duration.
Parameters of the secondary models predicting specific growth rate and lag phase are taken from \insertCite{Jia2020;textual}{qraLm}.
}
\note{
The work-to-be-done concept from \insertCite{Daelman2013;textual}{qraLm} was implemented
to keep track of the lapse of the lag phase or the remaining lag phase during multiple storage stages.
}
\examples{
dat <- Lot2LotGen(
  nLots = 100,
  sizeLot = 100,
  unitSize = 500,
  betaAlpha = 0.5112,
  betaBeta = 9.959,
  C0MeanLog = 1.023,
  C0SdLog = 0.3267,
  propVarInter = 0.7
)
N <- matrix(1:12, ncol = 4, nrow = 3)
dat <- list(N = N, unitSize = 500)

# Example of the first stage growth is taking place
aLM <- 0.0581 # parameter for mu (h-1)
Tmin <- 1.3 # parameter for mu (h-1)
ALM <- 0.84 # parameter for LPD (h)
mLM <- 1.11 # parameter for LPD (h)
MPD <- 9.2 # log10CFU/g # value taken from Jia et al. (2020)
time <- 11
Temp <- 2

first <- sfGrowthLPD(dat,
  MPD = MPD,
  Temp = 15,
  time = 11
)

# Example with previous stage where growth took place
second <- sfGrowthLPD(dat,
  MPD = MPD,
  Temp = 23,
  time = 11,
  workDone = first$workDone
)

# On matrix
third <- sfGrowthLPD(dat,
  MPD = MPD,
  Temp = c(10, 25, 30),
  time = 11
)
}
\references{
\insertRef{Daelman2013}{qraLm}

\insertRef{Jia2020}{qraLm}
}
\author{
Regis Pouillot \email{rpouillot.work@gmail.com} and Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{holding-time}
\keyword{lag-phase}
\keyword{salmon}
\keyword{transportation}
