% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfMaceration.R
\name{sfMaceration}
\alias{sfMaceration}
\title{Reduction of \emph{L. monocytogenes} due to maceration of fish fillets with gravlax curing agents}
\usage{
sfMaceration(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  rGravadMean = 0.7,
  rGravadSd = 0.283
)
}
\arguments{
\item{data}{See \code{\link[=Lot2LotGen]{Lot2LotGen()}} function.}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{rGravadMean}{(\code{log10}) mean of the normal distribution about the log10 reduction in \emph{L. monocytogenes} in brine-injected
fish fillets due to smoking and 18-24-h maturation (scalar).}

\item{rGravadSd}{(\code{log10}) standard deviation of the normal distribution about the log10 reduction in \emph{L. monocytogenes}
in brine-injected fish fillets due to smoking and 18-24-h maturation  (scalar).}
}
\value{
A list of two elements:
\describe{
\item{N}{(CFU) A matrix of size nLots lots by sizeLot units containing the numbers of \emph{L. monocytogenes}
in fish fillets after macerating with gravlax ingredients}
\item{pSurvSmoking}{Probability of a microbial cell to survive the maceration treatment (vector).}
}
}
\description{
The function \code{\link[=sfMaceration]{sfMaceration()}} describes the effect of macerating the fish fillets with lemon juice, salt, sugar,
black pepper and dill during 72 hours at 4 ºC. Lot-specific log10 reduction values are sampled from a normal distribution.
}
\note{
The suggested parameters \code{rGravadMean=0.7} and \code{rGravadSd=0.283} defining the normal distribution about the
variability in the log10 reduction in \emph{L. monocytogenes} in fish with gravlax ingredients were obtained combining the results
from \insertCite{Lopes2023;textual}{qraLm} and \insertCite{Neunlist2005;textual}{qraLm}.
}
\examples{
rGravadMean=0.7
rGravadSd=0.283
dat <- Lot2LotGen(
                  nLots = 50,
                  sizeLot = 100,
                  unitSize = 500,
                  betaAlpha = 0.5112,
                  betaBeta = 9.959,
                  C0MeanLog = 1.023,
                  C0SdLog = 0.3267,
                  propVarInter = 0.7
                  )
gravadfish <- sfMaceration(dat, 
                           rGravadMean=rGravadMean,
                           rGravadSd=rGravadSd)
hist(gravadfish$N)

}
\references{
\insertRef{Lopes2023}{qraLm}
\insertRef{extraDistr}{qraLm}
\insertRef{Neunlist2005}{qraLm}
\insertRef{stats}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{curing}
\keyword{effect}
\keyword{log-reduction}
\keyword{maturation}
\keyword{smoke}
