% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfMejlholmDalgaard.R
\name{sfMejlholmDalgaard}
\alias{sfMejlholmDalgaard}
\title{Secondary growth rate model of \emph{L. monocytogenes} in RTE seafood}
\usage{
sfMejlholmDalgaard(
  Temp,
  aw = NULL,
  NaCl = 0,
  pH = 7,
  P = 0,
  CO2equi = 0,
  NIT = 0,
  aaWph = 0,
  baWph = 0,
  caWph = 0,
  daWph = 0,
  laWph = 0,
  saWph = 0,
  Tmin = -2.83,
  Tref = 25,
  awmin = 0.923,
  pHmin = 4.97,
  pheMax = 32,
  NITmax = 350,
  CO2max = 3140,
  micLACu = 3.79,
  micDACu = 4.8,
  micAACu = 10.3,
  micBACu = 0.349,
  micCACu = 2.119,
  micSACu = 1.896,
  mumaxref = 0.419
)
}
\arguments{
\item{Temp}{(\eqn{^\circ C}) Storage temperature of the RTE product}

\item{aw}{Water activity of the RTE product. If NULL, it will evaluate from NaCl}

\item{NaCl}{(\%) Salt content of the RTE product.}

\item{pH}{pH of the RTE product.}

\item{P}{(ppm) Phenol concentration in the RTE product.}

\item{CO2equi}{(proportion) CO2 content in atmosphere in the RTE product (e.g. 0.25).}

\item{NIT}{(ppm) Nitrites concentration in the RTE product.}

\item{aaWph}{(ppm) Acetic acid concentration in the RTE product.}

\item{baWph}{(ppm) Benzoic acid concentration in the RTE product.}

\item{caWph}{(ppm) Citric acid concentration in the RTE product.}

\item{daWph}{(ppm) Diacetate concentration in the RTE product.}

\item{laWph}{(ppm) Lactic acid concentration in the RTE product.}

\item{saWph}{(ppm) Sorbic acid concentration in the RTE product.}

\item{Tmin}{(\eqn{^\circ C}) Minimum temperature for growth.}

\item{Tref}{(\eqn{^\circ C}) Reference temperature.}

\item{awmin}{Minimum water activity for growth.}

\item{pHmin}{Minimum pH for growth.}

\item{pheMax}{(ppm) MIC for phenols.}

\item{NITmax}{(ppm) MIC for nitrites.}

\item{CO2max}{(ppm) MIC for \code{CO_2}  .}

\item{micLACu}{(mM) MIC for undissociated lactic acid.}

\item{micDACu}{(mM) MIC for undissociated diacetate.}

\item{micAACu}{(mM) MIC for undissociated acetic acid.}

\item{micBACu}{(mM) MIC for undissociated benzoic acid.}

\item{micCACu}{(mM) MIC for undissociated citric acid.}

\item{micSACu}{(mM) MIC for undissociated sorbic acid.}

\item{mumaxref}{(1/h) Maximum growth rate at the reference temperature \code{Tref}.}
}
\value{
mumax (\code{1/h}) Specific growth rate of \emph{L. monocytogenes} in RTE seafood at the given intrinsic and extrinsic characteristics (scalar, vector or matrix).
}
\description{
The function \code{\link[=sfMejlholmDalgaard]{sfMejlholmDalgaard()}} estimates the growth rate of \emph{L. monocytogenes} in RTE seafood, based on the
predictive secondary model published by \insertCite{Mejlholm2009;textual}{qraLm}, and later validated in RTE foods including
seafood (\insertCite{Mejlholm2010}{qraLm}).
}
\note{
The parameters of the growth rate model are set by default to those fitted and validated by \insertCite{Mejlholm2010;textual}{qraLm}.
This model can also be found in the user friendly software FSSP: \href{http://fssp.food.dtu.dk/Help/Listeria/Lm-LAB/lm-lab.htm}{}.
}
\examples{
sfMejlholmDalgaard(Temp = c(-3, 15, 12), aw = 0.98, P = 3, saWph = 6000)

}
\references{
\insertRef{Mejlholm2009}{qraLm}

\insertRef{Mejlholm2010}{qraLm}
}
\author{
Laurent Guillier
}
\keyword{growth}
\keyword{lightly-preserved}
\keyword{model}
\keyword{seafood}
\keyword{secondary}
