% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfMejlholmDalgaardLAB.R
\name{sfMejlholmDalgaardLAB}
\alias{sfMejlholmDalgaardLAB}
\title{Secondary growth rate model of Lactic Acid Bacteria (LAB) in RTE seafood}
\usage{
sfMejlholmDalgaardLAB(
  Temp,
  aw = NULL,
  NaCl = 0,
  pH = 7,
  P = 0,
  CO2equi = 0,
  NIT = 0,
  aaWph = 0,
  baWph = 0,
  caWph = 0,
  daWph = 0,
  laWph = 0,
  saWph = 0,
  Tmin = -5.25,
  Tref = 25,
  awmin = 0.928,
  pHmin = 4.24,
  pheMax = 40.3,
  NITmax = 2780,
  CO2max = 6691,
  micLACu = 12,
  micDACu = 33.3,
  micAACu = 10.3,
  micBACu = 1.51,
  micCACu = 10.3,
  micSACu = 12.6,
  mumaxref = 0.583
)
}
\arguments{
\item{Temp}{(\eqn{^\circ C}) Storage temperature of the RTE product}

\item{aw}{Water activity of the RTE product. If NULL, it will evaluate from NaCl}

\item{NaCl}{(\%) Salt content of the RTE product.}

\item{pH}{pH of the RTE product.}

\item{P}{(ppm) Phenol concentration in the RTE product.}

\item{CO2equi}{(proportion) CO2 content in atmosphere in the RTE product (e.g. 0.25).}

\item{NIT}{(ppm) Nitrites concentration in the RTE product.}

\item{aaWph}{(ppm) Acetic acid concentration in the RTE product.}

\item{baWph}{(ppm) Benzoic acid concentration in the RTE product.}

\item{caWph}{(ppm) Citric acid concentration in the RTE product.}

\item{daWph}{(ppm) Diacetate concentration in the RTE product.}

\item{laWph}{(ppm) Lactic acid concentration in the RTE product.}

\item{saWph}{(ppm) Sorbic acid concentration in the RTE product.}

\item{Tmin}{(\eqn{^\circ C}) Minimum temperature for growth.}

\item{Tref}{(\eqn{^\circ C}) Reference temperature.}

\item{awmin}{Minimum water activity for growth.}

\item{pHmin}{Minimum pH for growth.}

\item{pheMax}{(ppm) MIC for phenols.}

\item{NITmax}{(ppm) MIC for nitrites.}

\item{CO2max}{(ppm) MIC for \verb{CO_ 2}.}

\item{micLACu}{(mM) MIC for undissociated lactic acid.}

\item{micDACu}{(mM) MIC for undissociated diacetate.}

\item{micAACu}{(mM) MIC for undissociated acetic acid.}

\item{micBACu}{(mM) MIC for undissociated benzoic acid.}

\item{micCACu}{(mM) MIC for undissociated citric acid.}

\item{micSACu}{(mM) MIC for undissociated sorbic acid.}

\item{mumaxref}{(\code{1/h}) Maximum growth rate at the reference temperature \code{Tref}.}
}
\value{
mumax (\code{1/h}) Specific growth rate of LAB in RTE seafood at the
given intrinsic and extrinsic characteristics (scalar, vector or matrix).
}
\description{
The function \code{\link[=sfMejlholmDalgaard]{sfMejlholmDalgaard()}} estimates the growth rate of LAB in RTE seafood, based on the
predictive secondary model published by \insertCite{Mejlholm2013;textual}{qraLm}, and later validated in RTE foods including
seafood (\insertCite{Mejlholm2015}{qraLm}).
}
\note{
The parameters of the growth rate model are set by default to those fitted and
validated by \insertCite{Mejlholm2015;textual}{qraLm}.
This model can also be found in the user friendly
software FSSP: \href{http://fssp.food.dtu.dk/Help/Listeria/Lm-LAB/lm-lab.htm}{}.
}
\examples{
# Reproducing screenshot value of FSSP (Fig. 1 red curve)
sfMejlholmDalgaardLAB(Temp = 5, aw = NULL, NaCl = 3.0, 
pH = 6.2, P = 5, laWph = 8000)
sfMejlholmDalgaardLAB(Temp = rnorm(10, 5, 0.2), 
                      aw = NULL, NaCl = 3.0, 
                      pH = 6.2, P = 5, laWph = 8000)

}
\references{
\insertRef{Mejlholm2013}{qraLm}

\insertRef{Mejlholm2015}{qraLm}
}
\author{
Laurent Guillier
}
\keyword{growth}
\keyword{lightly-preserved}
\keyword{model}
\keyword{psychrotrophic}
\keyword{seafood}
\keyword{secondary}
