% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfPortioning.R
\name{sfPortioning}
\alias{sfPortioning}
\title{Portioning of the contents of a seafood pack at home}
\usage{
sfPortioning(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  unitSize = NULL,
  servingSize,
  bPortSF
)
}
\arguments{
\item{data}{a list with a minimum element:
\describe{
\item{N}{(CFU) A matrix of size nLots lots by sizeLot pack units  containing the numbers of \emph{L. monocytogenes} per pack before handling at home.}
}}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{unitSize}{(g) is the weight of a pack of RTE seafood (scalar).}

\item{servingSize}{(g) is the portion taken from a pack, which will later equals to the serving size (scalar or vector).}

\item{bPortSF}{dispersion factor of cells within the package (scalar or vector).}
}
\value{
the data object with modified:
\describe{
\item{N}{(CFU) A matrix of size nLots lots by sizeLot units containing the numbers of \emph{L. monocytogenes} in the portions of RTE seafood.}
}
}
\description{
The \code{\link[=sfPortioning]{sfPortioning()}} function represents the portioning of a pack of RTE seafood into a smaller unit to be consumed. It is assumed that the microbial cells present in a contaminated pack
are distributed into servings following a beta-binomial distribution, although the algorithm only retains one portion per pack (and not all the portions that can be obtained from a pack).
The dispersion factor \code{bPortSF} represents the extent of cell clustering in the RTE seafood within the package.
}
\note{
A dispersion factor \code{bPortSF=1} represents moderate clustering of cells (\insertCite{Nauta2005}{qraLm}) in the RTE seafood within the package.
The serving size \code{servingSize} should be provided by the user and/or could be tested in scenarios.
}
\examples{
nLots <- 1000
sizeLot <- 500
dat <- list(
  N = matrix(rpois(sizeLot * nLots, 10),
    nrow = nLots, ncol = sizeLot
  ), P = 0.16,
  ProbUnitPos = rep(0.16, nLots)
)
Nf <- sfPortioning(dat, servingSize = 150, unitSize = 500, bPortSF = 1)
str(Nf)
hist(Nf$N) # histogram of microbial cells in contaminated servings

}
\references{
\insertRef{Nauta2005}{qraLm}

\insertRef{stats}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt} and Regis Pouillot \email{rpouillot.work@gmail.com}
}
\keyword{handling}
\keyword{portions}
\keyword{servings}
