% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfRawFishStorage.R
\name{sfRawFishStorage}
\alias{sfRawFishStorage}
\title{Storage of raw fish (whole or fillets) during processing}
\usage{
sfRawFishStorage(
  data = list(),
  unitSize = NULL,
  nLots = NULL,
  sizeLot = NULL,
  MPD = NULL,
  tempMin,
  tempMode,
  tempMax,
  timeMin,
  timeMode,
  timeMax
)
}
\arguments{
\item{data}{a list of four elements:
\describe{
\item{N}{(CFU) A matrix containing the numbers of \emph{L. monocytogenes} on whole fish before filleting, from contaminated lots;}
\item{ProbUnitPos}{Probability of individual lots being contaminated (vector);}
\item{P}{Prevalence of contaminated lots (scalar);}
\item{workDone}{The work done by bacteria.}
}}

\item{unitSize}{(g) Weight of a fish unit, whole fish or fish fillet (scalar, vector or matrix).}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{MPD}{(log10 CFU/g) Maximum population density of \emph{L. monocytogenes} on/in raw fish (scalar).}

\item{tempMin}{(\eqn{^\circ C}) Minimum value of the Pert distribution of variability about the holding temperature (scalar).}

\item{tempMode}{(\eqn{^\circ C}) Mode value of the Pert distribution of variability about the holding temperature (scalar).}

\item{tempMax}{(\eqn{^\circ C}) Maximum value of the Pert distribution of variability about the holding temperature (scalar).}

\item{timeMin}{(h) Minimum value of the Pert distribution of variability about the holding time (scalar).}

\item{timeMode}{(h) Mode value of the Pert distribution of variability about the holding time (scalar).}

\item{timeMax}{(h) Maximum value of the Pert distribution of variability about the holding time (scalar).}
}
\value{
A list of four elements:
\describe{
\item{N}{(CFU) A matrix containing the numbers of \emph{L. monocytogenes} on whole fish after the holding time, from contaminated lots;}
\item{ProbUnitPos}{Probability of individual lots being contaminated (vector);}
\item{P}{Prevalence of contaminated lots (scalar);}
\item{workDone}{Total work done during the previous stages including the current storage stage (vector).}
}
}
\description{
The function \code{\link[=sfRawFishStorage]{sfRawFishStorage()}} simulates the growth of \emph{L. monocytogenes} in whole fish before any short storage, such as
the holding time before filleting at primary processing, or in fish fillets during holding time before secondary processing; and is based on the
function \code{\link[=sfGrowthLPD]{sfGrowthLPD()}}. The algorithm is built upon two secondary models: one for lag-phase duration, and the other for growth rate,
both as functions of temperature, taken from \insertCite{@Jia2020;textual}{qraLm}. Log-linear growth is assumed until reaching
the maximum population density \code{MPD}. Pert distributions represent the lot-specific variability in holding time and temperature of raw fish.
}
\note{
The default value of \code{MPD=9.2} log10 CFU/g of \emph{L. monocytogenes} in raw fish was taken from \insertCite{@Jia2020;textual}{qraLm}.
Parameters for the holding time and temperature distributions should be defined by the user an/or tested in scenarios.
In addition to the final \emph{L. monocytogenes} numbers, the function also returns the lot-specific values of \code{workDone} so that the
\emph{L. monocytogenes} growth could be followed up in subsequent storage stages, if needed.
}
\examples{
tempMin <- -1.5
tempMode <- 0.5
tempMax <- 3
timeMin <- 0.5
timeMode <- 2
timeMax <- 6
MPD <- 9.2
nLots <- 1000
sizeLot <- 500
unitSize <- 1200
## No previous storage stage
dat <- list(
  N = matrix(
    rpois(nLots * sizeLot, 23),
    nLots,
    sizeLot
  ),
  ProbUnitPos = rep(0.15, nLots),
  P = 0.4,
  unitSize = unitSize
)
SecProc1 <- sfRawFishStorage(dat,
  unitSize,
  MPD = MPD,
  tempMin = tempMin,
  tempMode = tempMode,
  tempMax = tempMax,
  timeMin = timeMin,
  timeMode = timeMode,
  timeMax = timeMax
)
hist(SecProc1$N)
## Previous storage stage
dat <- list(
  N = matrix(rpois(nLots * sizeLot, 23), nLots, sizeLot),
  ProbUnitPos = rep(0.15, nLots), workDone = rep(0.5, nLots),
  P = 0.4,
  unitSize = unitSize
)
SecProc2 <- sfRawFishStorage(dat,
  unitSize = 1200,
  MPD = MPD,
  tempMin = tempMin,
  tempMode = tempMode,
  tempMax = tempMax,
  timeMin = timeMin,
  timeMode = timeMode,
  timeMax = timeMax
)
hist(SecProc2$N)

}
\references{
\insertRef{mc2d}{qraLm}

\insertRef{Jia2020}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{growth}
\keyword{lag-phase}
\keyword{logistics}
\keyword{stochastic}
