% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfSlicer.R
\name{sfSlicer}
\alias{sfSlicer}
\title{Cross-contamination during slicing of raw fish or processed fillets}
\usage{
sfSlicer(
  data = list(),
  nLots = NULL,
  sizeLot = NULL,
  wSlices,
  initSlicer = 10,
  aParamLoc = 0.07,
  aParamScale = 0.03,
  aParamMax = 0.5,
  eMean = -2.12,
  eSd = 0.85
)
}
\arguments{
\item{data}{a list of
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units containing the numbers of \emph{L. monocytogenes}
in/on the pre-slicing units;}
\item{\code{workDone}}{Work done is optional (scalar, vector or matrix).}
}}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{wSlices}{Weight of a slice that can be obtained from one unit (scalar).}

\item{initSlicer}{(\code{CFU}) Initial contamination of the slicer (scalar or vector). Default is 0.}

\item{aParamLoc}{Location parameter of the logistic distribution of \code{a}, the major parameter of the system (scalar).}

\item{aParamScale}{Scale parameter of the logistic distribution of \code{a}, the major parameter of the system (scalar).}

\item{aParamMax}{Maximum value used to truncate the logistic distribution of \code{a}, the major parameter of the  system (scalar).}

\item{eMean}{Mean of the log10 distribution of the \code{e} parameter (scalar)}

\item{eSd}{Standard deviation of the log10 distribution of the \code{e} parameter (scalar).}
}
\value{
A list of two elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size nLots lots by sizeLot units containing the numbers of \emph{L. monocytogenes}
in salted fish;}
\item{\code{workDone}}{Updated work done (scalar, vector or matrix).}
}
}
\description{
The function \code{\link[=sfSlicer]{sfSlicer()}} mimics the transfer of \emph{L. monocytogenes} during the slicing process of raw whole fish or
of processed fish fillets, using the compartmental model published in \insertCite{Hoelzer2012;textual}{qraLm}, which is
defined by two variability distributions: \code{a}, the transfer rate between slicer blade and product, and
\code{e}, the transfer rate from the original contamination to the slicing system. Transfer coefficients are sampled for every lot.
}
\note{
The parameters of the distributions about the slicer transfer coefficients \code{a} and \code{e} are set to the values
determined in \insertCite{Hoelzer2012;textual}{qraLm}; although the user can alter or update any of them.
The number of columns of the input matrix \code{N} is increased by a factor of \code{nbSlices}.
The code is written so that if a workDone object is element of \code{data}, it is updated as needed.
When this function is used for slicing processed fillets, \code{nbSlices} can be assumed to be 40 (slices per fillet),
according to \insertCite{Aarnisalo2007;textual}{qraLm}.
}
\examples{
nLots <- 1000
sizeLot <- 1000
N <- matrix(1E5, nc = nLots, nr = sizeLot)
data <- list(N = N, unitSize = 1300)

wSlices <- 20 # slice/fillet Aarnisalo 2007
Res <- sfSlicer(data, wSlices = wSlices, initSlicer = 0)
plot(log10(Res$N[1, ]), ylab = "log10 bacteria on each slices")
}
\references{
\insertRef{Aarnisalo2007}{qraLm}

\insertRef{extraDistr}{qraLm}

\insertRef{Hoelzer2012}{qraLm}

\insertRef{stats}{qraLm}
}
\author{
Regis Pouillot \email{rpouillot.work@gmail.com}
}
\keyword{coefficient}
\keyword{contamination}
\keyword{transfer}
