% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfSmoking.R
\name{sfSmoking}
\alias{sfSmoking}
\title{Reduction of \emph{L. monocytogenes} due to smoking brined or dry-salted fish}
\usage{
sfSmoking(
  data = list(),
  saltingType = NULL,
  nLots = NULL,
  sizeLot = NULL,
  rBrineMean = 0.871,
  rBrineSd = 0.807,
  rDrysaltMean = 1.093,
  rDrysaltSd = 0.532
)
}
\arguments{
\item{data}{See \code{\link[=Lot2LotGen]{Lot2LotGen()}} function.}

\item{saltingType}{Salting method employed in each of the lots: \code{brined} or \code{salted} (vector).}

\item{nLots}{Number of lots sampled or size of the Monte Carlo simulation (scalar).}

\item{sizeLot}{Number of units or portions produced in a lot (scalar).}

\item{rBrineMean}{(\code{log10}) mean of the normal distribution about the log10 reduction in \emph{L. monocytogenes} in brine-injected
fish fillets due to smoking and 18-24-h maturation (scalar).}

\item{rBrineSd}{(\code{log10}) standard deviation of the normal distribution about the log10 reduction in \emph{L. monocytogenes}
in brine-injected fish fillets due to smoking and 18-24-h maturation  (scalar).}

\item{rDrysaltMean}{(\code{log10}) mean of the normal distribution about the log10 reduction in \emph{L. monocytogenes} in dry-salted
fish fillets due to smoking and 18-24-h maturation (scalar).}

\item{rDrysaltSd}{(\code{log10}) standard deviation of the normal distribution about the log10 reduction in \emph{L. monocytogenes}
in dry-salted fish fillets due to smoking and 18-24-h maturation (scalar).}
}
\value{
A list of two elements:
\describe{
\item{N}{(CFU) A matrix of size nLots lots by sizeLot units containing the numbers of \emph{L. monocytogenes}
in salted fish fillets after smoking and 18-24-h maturation}
\item{pSurvSmoking}{Probability of a microbial cell to survive the smoking treatment (vector).}
}
}
\description{
The function \code{\link[=sfSmoking]{sfSmoking()}} describes the combined effect of smoking brined or salted fish fillets and maturing
for 18-24 hours. Based on the literature, different reduction factors of \emph{L. monocytogenes} are applied to
brine-injected fish fillets and dry-salted fillets. Lot-specific log10 reduction values are sampled from normal
distributions, attending to the type of salting.
}
\note{
The suggested parameters \code{rBrineMean=0.871} and \code{rBrineSd=0.807} defining the normal distribution about the
variability in the log10 reduction in \emph{L. monocytogenes} in brined-injected fish were modelled using data extracted from
\insertCite{Eklund1995;textual}{qraLm} and \insertCite{Porsby2008;textual}{qraLm}; where these authors inoculated \emph{L. monocytogenes}
in smoked salmon through brine injection; submitted samples to cycles of cold-smoking between 6-8 hours; and determined the microbial
concentrations after 18-24 hours maturation. The suggested parameters \code{rDrysaltMean=1.093} and \code{rDrysaltSd=0.532} characterising
the normal distribution about the variability in the log10 reduction in \emph{L. monocytogenes} in dry-salted fish were modelled using data from
\insertCite{Eklund1995;textual}{qraLm}, \insertCite{Neunlist2005;textual}{qraLm} and \insertCite{Porsby2008;textual}{qraLm}, who performed
experiments inoculating inoculated \emph{L. monocytogenes} on the surface of dry-salted salmon, and quantified the pathogen concentrations after
18-24 hours maturation.
}
\examples{
dat <- Lot2LotGen(
                  nLots = 50,
                  sizeLot = 100,
                  unitSize = 500,
                  betaAlpha = 0.5112,
                  betaBeta = 9.959,
                  C0MeanLog = 1.023,
                  C0SdLog = 0.3267,
                  propVarInter = 0.7
                  )
smokedfish <- sfSmoking(dat,
                        rBrineMean = 0.871, 
                        rBrineSd = 0.807,
                        rDrysaltMean = 1.093, 
                        rDrysaltSd = 0.532,
                        saltingType = "salted"
                        )
hist(smokedfish$N)
}
\references{
\insertRef{Eklund1995}{qraLm}
\insertRef{extraDistr}{qraLm}
\insertRef{Neunlist2005}{qraLm}
\insertRef{Porsby2008}{qraLm}
\insertRef{stats}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{curing}
\keyword{effect}
\keyword{log-reduction}
\keyword{maturation}
\keyword{smoke}
