% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryLot.qraLm.R
\name{summaryLot.qraLm}
\alias{summaryLot.qraLm}
\title{summaryLot Generic print function to print the between lots MC summary statistics}
\usage{
summaryLot.qraLm(x, ...)
}
\arguments{
\item{x}{qraLm object. See \code{\link[=Lot2LotGen]{Lot2LotGen()}}}

\item{...}{optional plot parameters passed to the plot function}
}
\description{
Print summary MC results at Lot level
}
\examples{
prod <- Lot2LotGen(
                   nLots        = 1000,
                   sizeLot      = 1000,
                   unitSize     = 500,
                   betaAlpha    = 0.5112,
                   betaBeta     = 9.959,
                   C0MeanLog    = 1.023,
                   C0SdLog      = 0.3267,
                   propVarInter = 0.7,
                   Poisson      = FALSE
)

summaryLot.qraLm(prod)

}
\author{
Vasco Cadavez
}
