% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryRiskLot.qraLm.R,
%   R/summaryRiskServings.qraLm.R
\name{summaryRisk.qraLm}
\alias{summaryRisk.qraLm}
\title{summaryRisk generic function to print the risk summary statistics}
\usage{
summaryRisk.qraLm(x, ...)

summaryRisk.qraLm(x, ...)
}
\arguments{
\item{x}{qraLm object. See \code{\link[=Lot2LotGen]{Lot2LotGen()}}}

\item{...}{optional plot parameters passed to the plot function}
}
\description{
Print summary MC risk results per lot

Print summary MC risk results per contamiunated serving
}
\examples{
dat <- Lot2LotGen(
                  nLots = 500,
                  sizeLot = 500,
                  unitSize = 500,
                  betaAlpha = 0.5112,
                  betaBeta = 9.959,
                  C0MeanLog = 1.023,
                  C0SdLog = 0.3267,
                  propVarInter = 0.7
                  )
DRmodel = "JEMRA"
population = 2
res <- DRForModel(dat, 
                  model=DRmodel,
                  population = population)

summaryRisk.qraLm(res)

dat <- Lot2LotGen(
                  nLots = 500,
                  sizeLot = 500,
                  unitSize = 500,
                  betaAlpha = 0.5112,
                  betaBeta = 9.959,
                  C0MeanLog = 1.023,
                  C0SdLog = 0.3267,
                  propVarInter = 0.7
                  )
DRmodel = "JEMRA"
population = 2
res <- DRForModel(dat, 
                  model=DRmodel,
                  population = population)

summaryRisk.qraLm(res)

}
\author{
Vasco Cadavez
}
