% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryServings.qraLm.R
\name{summaryServings.qraLm}
\alias{summaryServings.qraLm}
\title{summaryServings Generic function to print the within lots (between servings) MC summary statistics}
\usage{
summaryServings.qraLm(x, ...)
}
\arguments{
\item{x}{qraLm object. See \code{\link[=Lot2LotGen]{Lot2LotGen()}}}

\item{...}{optional plot parameters passed to the plot function}
}
\description{
Print summary of MC results
}
\examples{
dat <- Lot2LotGen(
  nLots = 500,
  sizeLot = 500,
  unitSize = 500,
  betaAlpha = 0.5112,
  betaBeta = 9.959,
  C0MeanLog = 1.023,
  C0SdLog = 0.3267,
  propVarInter = 0.7
)

dat1 <- fvDefrost(dat,
  Temp = 8,
  time = 2,
  MPD = 8.00,
  Tmin = -1.18,
  meanEGR5 = 0.0117,
  sdEGR5 = 0.00816,
  servingSize = 50,
  pDefrost = 0
)

#summaryServings.qraLm(dat1)

}
\author{
Vasco Cadavez
}
