% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryUnits.qraLm.R
\name{summaryUnits.qraLm}
\alias{summaryUnits.qraLm}
\title{summaryUnits Generic function to print the within lots (between units) MC summary statistics}
\usage{
summaryUnits.qraLm(x, ...)
}
\arguments{
\item{x}{qraLm object. See \code{\link[=Lot2LotGen]{Lot2LotGen()}}}

\item{...}{optional plot parameters passed to the plot function}
}
\description{
Print summary MC units results
}
\examples{
dat <- Lot2LotGen(
  nLots = 500,
  sizeLot = 500,
  unitSize = 500,
  betaAlpha = 0.5112,
  betaBeta = 9.959,
  C0MeanLog = 1.023,
  C0SdLog = 0.3267,
  propVarInter = 0.7
)

summaryUnits.qraLm(dat)

}
\author{
Vasco Cadavez
}
