% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caGrowthBaranyi.R
\name{caGrowthBaranyi}
\alias{caGrowthBaranyi}
\title{Growth of \emph{L. monocytogenes} in cantaloupe flesh}
\usage{
caGrowthBaranyi(
  data = list(),
  Temp,
  MPD,
  unitSize = NULL,
  EGR5,
  Tmin,
  Q0,
  time
)
}
\arguments{
\item{data}{A list of four elements see \code{\link[=caPrimaryProduction]{caPrimaryProduction()}} function.}

\item{Temp}{(\eqn{^{\circ}C}) Temperature  of cantaloupe flesh at the logistic stage.}

\item{MPD}{(\eqn{log_{10}\ CFU/g}) Maximum population density of \emph{L. monocytogenes} in cantaloupe flesh.}

\item{unitSize}{(\code{g}) Weight of a pack of cantaloupe dices.}

\item{EGR5}{(\eqn{h^{-1}}) Growth rate of \emph{L. monocytogenes} in cantaloupe flesh at 5 \eqn{^{\circ}C}.}

\item{Tmin}{(\eqn{^{\circ}C}) Nominal minimum temperature for growth of \emph{L. monocytogenes}in cantaloupe flesh.}

\item{Q0}{Initial parameter \code{Q} of the Baranyi and Roberts' model related to lag phase.}

\item{time}{(\code{h}) Time of the logistic stage.}
}
\value{
A list of two elements:
\describe{
\item{\code{Nt}}{Numbers (\code{CFU}) of \emph{L. monocytogenes} in cantaloupe flesh at the end of the stage (scalar, vector or matrix);}
\item{\code{lnQt}}{natural logarithm of the lag-phase parameter \code{Q} at the end of the stage (scalar, vector or matrix).}
}
}
\description{
The function \code{\link[=caGrowthBaranyi]{caGrowthBaranyi()}} estimates the numbers of \emph{L. monocytogenes} in cantaloupe flesh (\emph{i.e.}, dices, slices)
kept at a constant temperature \code{Temp} at the end of the logistics stage, \code{time}. The growth rate at a given temperature
is calculated from a square-root type model defined by the minimum temperature for growth (\code{Tmin}) and the exponential growth rate
at 5 \eqn{^{\circ}C} (\code{EGR5}). The primary growth model with lag phase of \insertCite{Baranyi1994;textual}{qraLm} is used to
estimate the numbers of \emph{L. monocytogenes} at \code{time}. The value of the physiological state of cells,
\code{q0}, must be provided. The return of the natural logarithm of \code{Q}
at \code{time} enables the reuse of this function for multiple transport and storage stages.
}
\note{
The parameters \code{Tmin} and \code{EGR5} were obtained by fitting a square-root model to growth data
of \emph{L. monocytogenes} in cantaloupe flesh at temperature below 30 \eqn{^\circ C}, extracted
from \insertCite{Danyluk2014;textual}{qraLm}, \insertCite{Fang2013;textual}{qraLm},
\insertCite{Farber1998;textual}{qraLm}, \insertCite{Guzel2017;textual}{qraLm}, \insertCite{Hong2014;textual}{qraLm},
\insertCite{Moreira2019;textual}{qraLm}, \insertCite{Patil2017;textual}{qraLm} and \insertCite{Ukuku2012;textual}{qraLm}.
\eqn{Tmin = -2.0196\ ^\circ C} (\eqn{standard\ error = 0.576\ ^\circ C}) and \eqn{EGR5 = 0.035573\ h^{-1}}
(\eqn{standard\ error = 0.004\ h^{-1}}). A zero is returned in the output elements when \eqn{N=0}.
}
\examples{

# Cells with a low Q0
dat <- Lot2LotGen(
  nLots = 50,
  sizeLot = 100,
  unitSize = 500,
  betaAlpha = 0.5112,
  betaBeta = 9.959,
  C0MeanLog = 1.023,
  C0SdLog = 0.3267,
  propVarInter = 0.7
)
Low <- caGrowthBaranyi(dat,
  Temp = 12,
  MPD = 8,
  unitSize = 250,
  EGR5 = 0.035573,
  Tmin = -2.0196,
  Q0 = 0.01,
  time = 5
)
hist(Low$Nt)
# Cells with a high Q0
High <- caGrowthBaranyi(dat,
  Temp = 12,
  MPD = 8,
  unitSize = 250,
  EGR5 = 0.035573,
  Tmin = -2.0196,
  Q0 = 3,
  time = 5
)
hist(High$Nt)
}
\references{
\insertRef{stats}{qraLm}

\insertRef{Danyluk2014}{qraLm}

\insertRef{Fang2013}{qraLm}

\insertRef{Farber1998}{qraLm}

\insertRef{Guzel2017}{qraLm}

\insertRef{Hong2014}{qraLm}

\insertRef{Moreira2019}{qraLm}

\insertRef{Patil2017}{qraLm}

\insertRef{Ukuku2002}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{Baranyi}
\keyword{RTE}
\keyword{dices}
\keyword{melon}
\keyword{pre-cut}
\keyword{slices}
