% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qraLm-package.R
\docType{package}
\name{qraLm-package}
\alias{qraLm}
\alias{qraLm-package}
\title{qraLm: Functions to develop quantitative risk assessment for Listeria monocytogenes in foods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

qraLm: A R package for quantitative risk assessment for Listeria monocytogenes in foods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vcadavez/qraLm}
  \item Report bugs at \url{https://github.com/vcadavez/qraLm/issues}
}

}
\author{
\strong{Maintainer}: Vasco Cadavez \email{vcadavez@ipb.pt} (\href{https://orcid.org/0000-0002-3077-7414}{ORCID})

Authors:
\itemize{
  \item Régis Pouillot \email{rpouillot.work@gmail.com} (\href{https://orcid.org/0000-0002-6107-5212}{ORCID})
  \item Laurent Guillier \email{Laurent.guillier@anses.fr} (\href{https://orcid.org/0000-0002-7867-2937}{ORCID})
  \item Moez Sanaa \email{sanaam@who.int} (\href{https://orcid.org/0000-0003-3576-0379}{ORCID})
  \item Ursula Gonzales-Barron \email{ubarron@ipb.pt} (\href{https://orcid.org/0000-0002-8462-9775}{ORCID})
}

}
\keyword{internal}
