% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfBrineORsaltCC.R
\name{sfBrineORsaltCC}
\alias{sfBrineORsaltCC}
\title{Cross-contamination of fish fillets during brine injection or dry-salting}
\usage{
sfBrineORsaltCC(
  data = list(),
  pBrine,
  pccBrine,
  volInjMin,
  volInjMode,
  volInjMax,
  concBrineMin,
  concBrineMode,
  concBrineMax,
  pccSmearing,
  nSurface,
  trSmearingMean,
  trSmearingSd
)
}
\arguments{
\item{data}{a list of
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size \code{nLots} lots by \code{sizeLot} units containing the numbers
of \emph{L. monocytogenes} in/on fish fillets;}
\item{\code{P}}{Mean prevalence of contaminated lots (scalar);}
\item{\code{ProbUnitPos}}{Probability of individual lots being contaminated (vector).}
}}

\item{pBrine}{Probability that a lot of fish fillets is salted by brine injection (as opposed to dry-salting) (scalar).}

\item{pccBrine}{Probability that the brine solution is contaminated with \emph{L. monocytogenes}  (scalar).}

\item{volInjMin}{(\code{ml}) minimum volume of brine solution injected in a fish fillet (scalar).}

\item{volInjMode}{(\code{ml}) most likely volume of brine solution injected in a fish fillet (scalar).}

\item{volInjMax}{(\code{ml}) maximum volume of brine solution injected in a fish fillet (scalar).}

\item{concBrineMin}{(\code{CFU/ml}) minimum concentration of \emph{L. monocytogenes} in contaminated brine solution (scalar).}

\item{concBrineMode}{(\code{CFU/ml}) most likely concentration of \emph{L. monocytogenes} in contaminated brine solution (scalar).}

\item{concBrineMax}{(\code{CFU/ml}) maximum concentration of \emph{L. monocytogenes} in contaminated brine solution (scalar).}

\item{pccSmearing}{Probability that cross-contamination with \emph{L. monocytogenes} occurs during dry salting, or smearing
the fillets with sugar/spices, through tables or other surfaces (scalar).}

\item{nSurface}{(CFU) Numbers of \emph{L. monocytogenes} cells on table or surfaces in contact with fish fillets while dry salting
(scalar or vector).}

\item{trSmearingMean}{Mean parameter of the normal distribution representing the variability in the log 10 of the transfer coefficient
of \emph{L. monocytogenes} cells from surfaces to fish fillets (scalar or vector).}

\item{trSmearingSd}{Standard deviation parameter of the normal distribution representing the variability in the log 10 of the transfer
coefficient of \emph{L. monocytogenes} from surfaces to fish fillets (scalar or vector).}
}
\value{
A list of three elements:
\describe{
\item{\code{N}}{(\code{CFU}) A matrix of size nLots lots by sizeLot units containing the numbers of \emph{L. monocytogenes} in salted fish;}
\item{\code{ProbUnitPos}}{Probability of individual lots being contaminated after salting (vector);}
\item{\code{P}}{Mean prevalence of contaminated lots after salting (scalar).}
}
}
\description{
The function \code{\link[=sfBrineORsaltCC]{sfBrineORsaltCC()}} simulates the potential internal or external cross-contamination of fish fillets during the process of salting, either by
brine injection or by dry-salting, respectively. The type of salting applied in a lot of fish fillets is mutually exclusive; so
the probability that a lot of fish fillets is salted via brine injection is \code{pBrine}, and the complement \code{1-Pcc_salting} is the
probability that the lot of fish fillets is treated by external dry-salting. Since this algorithm relies on two functions: \code{\link[=sfBriningCC]{sfBriningCC()}} and
\code{\link[=sfSmearingCC]{sfSmearingCC()}}, the arguments of both functions must be provided to \code{\link[=sfBrineORsaltCC]{sfBrineORsaltCC()}}.
}
\note{
The suggested value of \eqn{probCCDice\_brine=0.135} is taken from \insertCite{Gudbjornsdottir2005;textual}{qraLm} and
\insertCite{Gudbjornsdottir2004;textual}{qraLm}, who analysed the presence of \emph{L. monocytogenes} in brine solution, detecting
3 positive samples out of 14, and 2 positive samples out of 23, respectively. The suggested parameters \eqn{TR\_smearing\_mean=-0.29}
and \eqn{TR\_smearing\_sd=0.31} for the normal distribution about the variability in the log 10 of the transfer coefficient of \emph{L. monocytogenes} were taken
from \insertCite{Hoelzer2012;textual}{qraLm}, to represent cross-contamination from board to meat.
The parameters of the Pert distributions about mean volume of brine solution injected in a fish fillet (\code{volInjMin}, \code{volInjMode} and \code{volInjMax}),
and about mean concentration of \emph{L. monocytogenes} in brine (\code{concBrineMin}, \code{concBrineMode} and \code{concBrineMax}) must be defined by the user and/or
tested in scenarios. Similarly, the values of \code{pBrine}, \code{Pcc_salting} and \code{nSurface} must be defined by the user and/or assessed in scenarios.
}
\examples{

dat <- Lot2LotGen(
  nLots = 50,
  sizeLot = 100,
  unitSize = 500,
  betaAlpha = 0.5112,
  betaBeta = 9.959,
  C0MeanLog = 1.023,
  C0SdLog = 0.3267,
  propVarInter = 0.7
)
df <- sfBrineORsaltCC(dat,
                      pBrine = 0.2,
                      pccBrine = 0.13,
                      volInjMin = 0.5,
                      volInjMode = 1.2,
                      volInjMax = 4.5,
                      concBrineMin = 0,
                      concBrineMode = 3,
                      concBrineMax = 40,
                      pccSmearing = 0.05,
                      trSmearingMean = -0.29,
                      trSmearingSd = 0.31,
                      nSurface = 200)
 hist(df$N)
}
\references{
\insertRef{Gudbjornsdottir2004}{qraLm}
\insertRef{Gudbjornsdottir2005}{qraLm}
\insertRef{Hoelzer2012}{qraLm}
\insertRef{stats}{qraLm}
}
\author{
Ursula Gonzales-Barron \email{ubarron@ipb.pt}
}
\keyword{brine}
\keyword{contamination}
\keyword{external}
\keyword{injection}
\keyword{internal}
\keyword{salt}
\keyword{smearing}
\keyword{solution}
\keyword{surface}
\keyword{transfer}
